Trait vizia::animation::AnimExt

pub trait AnimExt: Copy + Sized {
    fn entity(self) -> Entity;

    fn play_animation(
        self,
        cx: &mut EventContext<'_>,
        animation: Animation
    ) -> Self { ... }
    fn is_animating(self, cx: &mut Context, animation: Animation) -> bool { ... }
}

Trait which provides methods for entities to manipulate linked animations

This trait is part of the prelude.

Required Methods

fn entity(self) -> Entity

Provided Methods

fn play_animation(self, cx: &mut EventContext<'_>, animation: Animation) -> Self

Play an animation on the entity.

Internally this generates an active animation and links the entity to it for each animated property.

Example

Create an animation which animates the left property from 0 to 100 pixels in 5 seconds and play the animation on an entity:

let animation_id = cx.add_animation(instant::Duration::from_secs(5))
    .add_keyframe(0.0, |keyframe| keyframe.set_left(Pixels(0.0)))
    .add_keyframe(1.0, |keyframe| keyframe.set_left(Pixels(100.0)))
    .build();

entity.play_animation(cx, animation_id);

fn is_animating(self, cx: &mut Context, animation: Animation) -> bool

Returns true if there is an active animation with the given id.

Example

let test = entity.is_animating(animation_id);