Struct vizia::cache::BoundingBox

pub struct BoundingBox {
    pub x: f32,
    pub y: f32,
    pub w: f32,
    pub h: f32,
}

Respresents an axis-aligned bounding box of an entity.

Fields
x: f32y: f32w: f32h: f32
Implementations

impl BoundingBox

pub fn from_min_max(min_x: f32, min_y: f32, max_x: f32, max_y: f32) -> BoundingBox

Construct a BoundingBox from checked minimum and maximum values.

pub fn left(&self) -> f32

Left side of bounds equivalent to x

pub fn top(&self) -> f32

Top of bounds equivalent to y

pub fn width(&self) -> f32

Bounds width equivalent to w

pub fn height(&self) -> f32

Bounds height equivalent to h

pub fn right(&self) -> f32

Right side of bounds

pub fn bottom(&self) -> f32

Bottom of bounds

pub fn center(&self) -> (f32, f32)

Horizontal and vertical center of bounds

pub fn center_left(&self) -> (f32, f32)

Left center of bounds

pub fn center_right(&self) -> (f32, f32)

Right center of bounds

pub fn center_top(&self) -> (f32, f32)

Top center of bounds

pub fn center_bottom(&self) -> (f32, f32)

Bottom center of bounds

pub fn bottom_left(&self) -> (f32, f32)

Bottom left point of bounds

pub fn bottom_right(&self) -> (f32, f32)

Bottom right point of bounds

pub fn top_left(&self) -> (f32, f32)

Top left point of bounds

pub fn top_right(&self) -> (f32, f32)

Top right point of bounds

pub fn shrink(&self, amount: f32) -> BoundingBox

Shrinks by some amount in both directions and returns a new BoundingBox.

pub fn shrink_horizontal(&self, amount: f32) -> BoundingBox

Shrinks by some amount horizontally and returns a new BoundingBox.

pub fn shrink_vertical(&self, amount: f32) -> BoundingBox

Shrinks by some amount vertically and returns a new BoundingBox.

pub fn expand(&self, amount: f32) -> BoundingBox

Expands by some amount in both directions and returns a new BoundingBox.

pub fn expand_horizontal(&self, amount: f32) -> BoundingBox

Expands by some amount horizontally and returns a new BoundingBox.

pub fn expand_vertical(&self, amount: f32) -> BoundingBox

Expands by some amount vertically and returns a new BoundingBox.

pub fn intersects(&self, other: &BoundingBox) -> bool

pub fn contains(&self, other: &BoundingBox) -> bool