Struct vizia::cache::CachedData

pub struct CachedData { /* private fields */ }

Stores data which can be cached between system runs.

When an event occurs or style data is changed systems run to determine the new state of the UI. The output of these systems can be cached so that not all of the systems need to run again.

Implementations

impl CachedData

pub fn add(&mut self, entity: Entity) -> Result<(), SparseSetError>

pub fn remove(&mut self, entity: Entity)

pub fn get_grid_row_max(&self, entity: Entity) -> f32

pub fn set_grid_row_max(&mut self, entity: Entity, value: f32)

pub fn get_grid_col_max(&self, entity: Entity) -> f32

pub fn set_grid_col_max(&mut self, entity: Entity, value: f32)

pub fn get_stack_child(&self, entity: Entity) -> (bool, bool)

pub fn get_bounds(&self, entity: Entity) -> BoundingBox

Returns the bounding box of the entity, determined by the layout system.

pub fn get_clip_region(&self, entity: Entity) -> BoundingBox

Returns the clip region of the entity.

This is the bounding box for which rendering of the widget will be clipped/cropped when outside of the bounds.

pub fn get_child_width_sum(&self, entity: Entity) -> f32

pub fn get_child_height_sum(&self, entity: Entity) -> f32

pub fn get_child_width_max(&self, entity: Entity) -> f32

pub fn get_child_height_max(&self, entity: Entity) -> f32

pub fn get_posx(&self, entity: Entity) -> f32

Returns the x position of the entity.

pub fn get_posy(&self, entity: Entity) -> f32

Returns the y position of the entity.

pub fn get_width(&self, entity: Entity) -> f32

Returns the width of the entity.

pub fn get_height(&self, entity: Entity) -> f32

Returns the height of the entity.

pub fn get_opacity(&self, entity: Entity) -> f32

Returns the opacity of the entity.

pub fn get_horizontal_free_space(&self, entity: Entity) -> f32

pub fn get_horizontal_stretch_sum(&self, entity: Entity) -> f32

pub fn get_vertical_free_space(&self, entity: Entity) -> f32

pub fn get_vertical_stretch_sum(&self, entity: Entity) -> f32

pub fn get_rotate(&self, entity: Entity) -> f32

pub fn get_translate(&self, entity: Entity) -> (f32, f32)

pub fn get_scale(&self, entity: Entity) -> f32

pub fn get_origin(&self, entity: Entity) -> (f32, f32)

pub fn get_transform(&self, entity: Entity) -> Transform2D

Returns the transform on the entity.

pub fn get_transform_mut(&mut self, entity: Entity) -> &mut Transform2D

pub fn set_stack_first_child(&mut self, entity: Entity, value: bool)

pub fn set_stack_last_child(&mut self, entity: Entity, value: bool)

pub fn set_horizontal_free_space(&mut self, entity: Entity, value: f32)

pub fn set_horizontal_stretch_sum(&mut self, entity: Entity, value: f32)

pub fn set_vertical_free_space(&mut self, entity: Entity, value: f32)

pub fn set_vertical_stretch_sum(&mut self, entity: Entity, value: f32)

pub fn set_space_right(&mut self, entity: Entity, val: f32)

pub fn set_clip_region(&mut self, entity: Entity, val: BoundingBox)

pub fn set_child_width_sum(&mut self, entity: Entity, val: f32)

pub fn set_child_height_sum(&mut self, entity: Entity, val: f32)

pub fn set_child_width_max(&mut self, entity: Entity, val: f32)

pub fn set_child_height_max(&mut self, entity: Entity, val: f32)

pub fn set_posx(&mut self, entity: Entity, val: f32)

pub fn set_posy(&mut self, entity: Entity, val: f32)

pub fn set_width(&mut self, entity: Entity, val: f32)

pub fn set_height(&mut self, entity: Entity, val: f32)

pub fn get_hoverability(&self, entity: Entity) -> bool

pub fn get_visibility(&self, entity: Entity) -> Visibility

pub fn get_display(&self, entity: Entity) -> Display

pub fn set_opacity(&mut self, entity: Entity, val: f32)