Struct vizia::context::backend::BackendContext

pub struct BackendContext<'a>(pub &'a mut Context);

Tuple Fields
0: &'a mut Context
Implementations

impl<'a> BackendContext<'a>

pub fn new(cx: &'a mut Context) -> BackendContext<'a>

pub fn views(
&mut self
) -> &mut HashMap<Entity, Box<dyn ViewHandler + 'static, Global>, BuildHasherDefault<FnvHasher>>

pub fn style(&mut self) -> &mut Style

pub fn cache(&mut self) -> &mut CachedData

pub fn modifiers(&mut self) -> &mut Modifiers

pub fn window_size(&mut self) -> &mut WindowSize

The window’s size in logical pixels, before user_scale_factor() gets applied to it. If this value changed during a frame then the window will be resized and a WindowEvent::GeometryChanged will be emitted.

pub fn user_scale_factor(&mut self) -> &mut f64

A scale factor used for uniformly scaling the window independently of any HiDPI scaling. window_size gets multplied with this factor to get the actual logical window size. If this changes during a frame, then the window will be resized at the end of the frame and a WindowEvent::GeometryChanged will be emitted. This can be initialized using [WindowDescription::user_scale_factor][crate::WindowDescription::user_scale_factor].

pub fn add_main_window(
&mut self,
window_description: &WindowDescription,
canvas: Canvas<OpenGl>,
dpi_factor: f32
)

pub fn environment(&self) -> &Environment

pub fn context(&mut self) -> &mut Context

pub fn remove_all_children(&mut self)

pub fn draw(&mut self)

pub fn load_images(&mut self)

pub fn set_current(&mut self, e: Entity)

Set the current entity. This is useful in user code when you’re performing black magic and want to trick other parts of the code into thinking you’re processing some other part of the tree.

pub fn set_text_config(&mut self, text_config: TextConfig)

Sets the default text configuration to use for text rendering.

pub fn with_current(&mut self, e: Entity, f: impl FnOnce(&mut Context))

Temporarily sets the current entity, calls the provided closure, and then resets the current entity back to previous.

pub fn set_event_proxy(&mut self, proxy: Box<dyn EventProxy + 'static, Global>)

You should not call this method unless you are writing a windowing backend, in which case you should consult the existing windowing backends for usage information.

pub fn set_clipboard_provider(
&mut self,
clipboard: Box<dyn ClipboardProvider + 'static, Global>
)

You should not call this method unless you are writing a windowing backend, in which case you should consult the existing windowing backends for usage information.

pub fn send_event(&mut self, event: Event)

Send an event with custom origin and propagation information.

pub fn has_queued_events(&self) -> bool

Check whether there are any events in the queue waiting for the next event dispatch cycle.

pub fn process_data_updates(&mut self)

For each binding or data observer, check if its data has changed, and if so, rerun its builder/body.

pub fn process_style_updates(&mut self)

pub fn process_animations(&mut self) -> bool

pub fn process_visual_updates(&mut self)

Massages the style system until everything is coherent

pub fn emit_origin<M>(&mut self, message: M)where
M: Send + Any,