Struct vizia::entity::Entity

pub struct Entity(_);

An entity is an identifier used to reference a view; to get/set properties in the context.

Rather than having widgets own their data, all state is stored in a single database and is stored and loaded using entities.

This type is part of the prelude.

Implementations

impl Entity

pub fn root() -> Entity

Creates a new root entity.

The root entity represents the main window and is always valid. It can be used to set properties on the primary window, such as background color, as well as sending events to the window such as Restyle and Redraw events.