Struct vizia::events::EventMeta

pub struct EventMeta {
    pub origin: Entity,
    pub target: Entity,
    pub propagation: Propagation,
    pub consumable: bool,
    pub order: i32,
    /* private fields */
}

The meta data of an Event.

Fields
origin: Entity

The entity that produced the event. Entity::null() for OS events or unspecified.

target: Entity

The entity the event should be sent to. Entity::null() to send to all entities.

propagation: Propagation

How the event propagates through the tree.

consumable: bool

Whether the event can be consumed

order: i32

Specifies an order index which is used to sort the event queue

Implementations

impl EventMeta

pub fn new() -> EventMeta

Creates a new event meta.

impl EventMeta

pub fn consume(&mut self)

Consumes the event to prevent it from continuing on its propagation path.