Trait vizia::events::ViewHandler

pub trait ViewHandler: Any {
    fn event(&mut self, cx: &mut EventContext<'_>, event: &mut Event);
    fn draw(&self, cx: &mut DrawContext<'_>, canvas: &mut Canvas<OpenGl>);
    fn as_any_ref(&self) -> &(dyn Any + 'static);
    fn as_any_mut(&mut self) -> &mut (dyn Any + 'static);

    fn element(&self) -> Option<&'static str> { ... }
}

Required Methods

fn event(&mut self, cx: &mut EventContext<'_>, event: &mut Event)

fn draw(&self, cx: &mut DrawContext<'_>, canvas: &mut Canvas<OpenGl>)

fn as_any_ref(&self) -> &(dyn Any + 'static)

fn as_any_mut(&mut self) -> &mut (dyn Any + 'static)

Provided Methods

fn element(&self) -> Option<&'static str>

Implementations

impl dyn ViewHandler + 'static

pub fn downcast_ref<T>(&self) -> Option<&T>where
T: Any,

Attempt to cast a view handler to an immutable reference to the specified type.

pub fn downcast_mut<T>(&mut self) -> Option<&mut T>where
T: Any,

Attempt to cast a view handler to a mutable reference to the specified type.