vizia::localization

Struct Localized

pub struct Localized { /* private fields */ }

A type which formats a localized message with any number of named arguments.

Implementations

impl Localized

pub fn new(key: &str) -> Localized

Creates a new Localized type with a given key.

The given key is used to retrieve a translation from a fluent bundle resource.

Example

Application::new(|cx|{
    Label::new(cx, Localized::new("key"));
})
.run();

pub fn map(self, mapping: impl Fn(&str) -> String + 'static) -> Localized

Sets a mapping function to apply to the translated text.

pub fn arg<L>(self, key: &str, lens: L) -> Localized
where L: Lens, <L as Lens>::Target: Into<FluentValue<'static>> + Data,

Add a variable argument binding to the Localized type.

Takes a key name and a lens to the value for the argument.

Example

Application::new(|cx|{
     
    AppData {
        value: 5,
    }.build(cx);

    Label::new(cx, Localized::new("key").arg("value", AppData::value));
})
.run();

pub fn arg_const<T>(self, key: &str, val: T) -> Localized
where T: Into<FluentValue<'static>> + Data,

Add a constant argument to the Localized type.

Takes a key name and a value for the argument.

Example

Application::new(|cx|{

    Label::new(cx, Localized::new("key").arg_const("value", 32));
})
.run();