Enum vizia::prelude::WindowEvent

pub enum WindowEvent {

    WindowClose,
    WindowResize(f32, f32),
    MouseDoubleClick(MouseButton),
    MouseTripleClick(MouseButton),
    MouseDown(MouseButton),
    MouseUp(MouseButton),
    Press {
        mouse: bool,
    },
    PressDown {
        mouse: bool,
    },
    MouseMove(f32, f32),
    MouseScroll(f32, f32),
    MouseOver,
    MouseOut,
    MouseEnter,
    MouseLeave,
    FocusIn,
    FocusOut,
    CharInput(char),
    KeyDown(Code, Option<Key>),
    KeyUp(Code, Option<Key>),
    SetCursor(CursorIcon),
    GrabCursor(bool),
    SetCursorPosition(u32, u32),
    SetTitle(String),
    SetSize(WindowSize),
    SetPosition(Position),
    SetMaxSize(Option<WindowSize>),
    SetMinSize(Option<WindowSize>),
    SetResizable(bool),
    SetMinimized(bool),
    SetMaximized(bool),
    SetVisible(bool),
    SetDecorations(bool),
    SetAlwaysOnTop(bool),
    MouseCaptureEvent,
    MouseCaptureOutEvent,
    GeometryChanged(GeometryChanged),
    Redraw,
    Restyle,
    Relayout,
    Debug(String),

}

Events generated by the application in response to OS events as well as events that can be used to set properties of the window.

This type is part of the prelude.

Variants

WindowClose

Emitted when a window is closed

WindowResize(f32, f32)

Emitted when a window is opened

MouseDoubleClick(MouseButton)

Emitted when a mouse button is double clicked

MouseTripleClick(MouseButton)

Emitted when a mouse button is triple clicked

MouseDown(MouseButton)

Emitted when a mouse button is pressed

MouseUp(MouseButton)

Emitted when a mouse button is released

Press

Fields

mouse: bool

Emitted when the primary mouse button or trigger key is pressed and then released on a view

PressDown

Fields

mouse: bool

Emitted when the primary mouse button or trigger key is pressed on a view

MouseMove(f32, f32)

Emitted when the mouse cursor is moved

MouseScroll(f32, f32)

Emitted when the mouse scroll wheel is scrolled

MouseOver

Emitted when the mouse cursor enters the bounding box of an entity

MouseOut

Emitted when the mouse cursor leaves the bounding box of an entity

MouseEnter

Emitted when the mouse cursor enters an entity or one of its descendants

MouseLeave

Emitted when the mouse cursor leaves an entity or one of its descendants

FocusIn

FocusOut

CharInput(char)

Emitted when a character is typed

KeyDown(Code, Option<Key>)

Emitted when a keyboard key is pressed

KeyUp(Code, Option<Key>)

Emitted when a keyboard key is released

SetCursor(CursorIcon)

Sets the mouse cursor icon

GrabCursor(bool)

Grabs the mouse cursor, preventing it from leaving the window

SetCursorPosition(u32, u32)

Sets the (x,y) position of the mouse cursor in window coordinates

SetTitle(String)

SetSize(WindowSize)

SetPosition(Position)

SetMaxSize(Option<WindowSize>)

SetMinSize(Option<WindowSize>)

SetResizable(bool)

SetMinimized(bool)

SetMaximized(bool)

SetVisible(bool)

SetDecorations(bool)

SetAlwaysOnTop(bool)

MouseCaptureEvent

Emitted when mouse events have been captured

MouseCaptureOutEvent

Emitted when mouse events have been released

GeometryChanged(GeometryChanged)

Emitted when an entity changes position or size (TODO: check if this includes margins + borders)

Redraw

Requests a redraw of the window contents

Restyle

Request a restyle

Relayout

Requests a relayout

Debug(String)

Prints the debug message to the console