Struct vizia::prelude::Animation

pub struct Animation(_);

An id used to reference style animations stored in context.

An animation id is returned by cx.add_animation() and can be used to configure animations as well as to play, pause, and stop animations on entities (see AnimExt).

This type is part of the prelude.