Struct vizia::prelude::Context

pub struct Context {
    pub views: HashMap<Entity, Box<dyn ViewHandler + 'static, Global>, BuildHasherDefault<FnvHasher>>,
    pub ignore_default_theme: bool,
    pub window_has_focus: bool,
    /* private fields */
}

The main storage and control object for a Vizia application.

This type is part of the prelude.

Fields
views: HashMap<Entity, Box<dyn ViewHandler + 'static, Global>, BuildHasherDefault<FnvHasher>>

TODO make this private when there’s no longer a need to mutate views after building

ignore_default_theme: boolwindow_has_focus: bool
Implementations

impl Context

pub fn new(window_size: WindowSize, user_scale_factor: f64) -> Context

pub fn current(&self) -> Entity

The “current” entity, generally the entity which is currently being built or the entity which is currently having an event dispatched to it.

pub fn set_current(&mut self, e: Entity)

Set the current entity. This is useful in user code when you’re performing black magic and want to trick other parts of the code into thinking you’re processing some other part of the tree.

pub fn with_current(&mut self, e: Entity, f: impl FnOnce(&mut Context))

Makes the above black magic more explicit

pub fn environment(&self) -> &Environment

pub fn window_size(&self) -> WindowSize

The window’s size in logical pixels, before user_scale_factor() gets applied to it. If this value changed during a frame then the window will be resized and a WindowEvent::GeometryChanged will be emitted.

pub fn user_scale_factor(&self) -> f64

A scale factor used for uniformly scaling the window independently of any HiDPI scaling. window_size gets multplied with this factor to get the actual logical window size. If this changes during a frame, then the window will be resized at the end of the frame and a WindowEvent::GeometryChanged will be emitted. This can be initialized using [WindowDescription::user_scale_factor][crate::WindowDescription::user_scale_factor].

pub fn needs_redraw(&mut self)

Mark the application as needing to rerun the draw method

pub fn needs_restyle(&mut self)

Mark the application as needing to recompute view styles

pub fn needs_relayout(&mut self)

Mark the application as needing to rerun layout computations

pub fn focus_with_visibility(&mut self, focus_visible: bool)

Sets application focus to the current entity with the specified focus visiblity

pub fn focus(&mut self)

Sets application focus to the current entity using the previous focus visibility

pub fn set_selected(&mut self, flag: bool)

Sets the checked flag of the current entity

pub fn remove(&mut self, entity: Entity)

pub fn add_listener<F, W>(&mut self, listener: F)where
W: View,
F: 'static + Fn(&mut W, &mut EventContext<'_>, &mut Event),

Check whether there are any events in the queue waiting for the next event dispatch cycle. Add a listener to an entity.

A listener can be used to handle events which would not normally propagate to the entity. For example, mouse events when a different entity has captured them. Useful for things like closing a popup when clicking outside of its bounding box.

pub fn add_global_listener<F>(&mut self, listener: F)where
F: 'static + Fn(&mut EventContext<'_>, &mut Event),

Adds a global listener to the application.

Global listeners have the first opportunity to handle every event that is sent in an application. They will never be removed. If you need a listener tied to the lifetime of a view, use add_listener.

pub fn add_fonts_mem(&mut self, data: &[&[u8]])

Add a font from memory to the application.

pub fn set_default_font(&mut self, names: &[&str])

Sets the global default font for the application.

pub fn add_theme(&mut self, theme: &str)

pub fn remove_user_themes(&mut self)

pub fn add_stylesheet(&mut self, path: impl AsRef<Path>) -> Result<(), Error>

pub fn add_animation(&mut self, duration: Duration) -> AnimationBuilder<'_>

Adds a new property animation returning an animation builder

Example

Create an animation which animates the left property from 0 to 100 pixels in 5 seconds and play the animation on an entity:

let animation_id = cx.add_animation(instant::Duration::from_secs(5))
    .add_keyframe(0.0, |keyframe| keyframe.set_left(Pixels(0.0)))
    .add_keyframe(1.0, |keyframe| keyframe.set_left(Pixels(100.0)))
    .build();

pub fn set_image_loader<F>(&mut self, loader: F)where
F: 'static + Fn(&mut ResourceContext<'_>, &str),

pub fn add_translation(&mut self, lang: LanguageIdentifier, ftl: String)

pub fn spawn<F>(&self, target: F)where
F: 'static + Send + FnOnce(&mut ContextProxy),

pub fn resolve_entity_identifier(&self, identity: &str) -> Option<Entity>

Finds the entity that identifier identifies