Struct vizia::prelude::EventContext

pub struct EventContext<'a> {
    pub cache: &'a CachedData,
    pub draw_cache: &'a DrawCache,
    pub tree: &'a Tree<Entity>,
    pub views: &'a HashMap<Entity, Box<dyn ViewHandler + 'static, Global>, BuildHasherDefault<FnvHasher>>,
    pub resource_manager: &'a ResourceManager,
    pub text_context: &'a TextContext,
    pub modifiers: &'a Modifiers,
    pub mouse: &'a MouseState<Entity>,
    /* private fields */
}

Fields
cache: &'a CachedDatadraw_cache: &'a DrawCachetree: &'a Tree<Entity>views: &'a HashMap<Entity, Box<dyn ViewHandler + 'static, Global>, BuildHasherDefault<FnvHasher>>resource_manager: &'a ResourceManagertext_context: &'a TextContextmodifiers: &'a Modifiersmouse: &'a MouseState<Entity>
Implementations

impl<'a> EventContext<'a>

pub fn new(cx: &'a mut Context) -> EventContext<'a>

pub fn resolve_entity_identifier(&self, identity: &str) -> Option<Entity>

Finds the entity that identifier identifies

pub fn current(&self) -> Entity

pub fn emit<M>(&mut self, message: M) where
M: Any + Send,

Send an event containing a message up the tree from the current entity.

pub fn emit_to<M>(&mut self, target: Entity, message: M) where
M: Any + Send,

Send an event containing a message directly to a specified entity.

pub fn emit_custom(&mut self, event: Event)

Send an event with custom origin and propagation information.

pub fn add_listener<F, W>(&mut self, listener: F) where
W: View,
F: 'static + Fn(&mut W, &mut EventContext<'_>, &mut Event),

Add a listener to an entity.

A listener can be used to handle events which would not normally propagate to the entity. For example, mouse events when a different entity has captured them. Useful for things like closing a popup when clicking outside of its bounding box.

pub fn set_active(&mut self, active: bool)

Set the active state for the current entity.

pub fn capture(&mut self)

Capture mouse input for the current entity.

pub fn release(&mut self)

Release mouse input capture for current entity.

pub fn focus_with_visibility(&mut self, focus_visible: bool)

Sets application focus to the current entity with the specified focus visiblity

pub fn focus(&mut self)

Sets application focus to the current entity using the previous focus visibility.

pub fn hovered(&self) -> Entity

Return the currently hovered entity.

pub fn focused(&self) -> Entity

Return the currently focused entity.

pub fn is_disabled(&self) -> bool

Returns true if the current entity is disabled.

pub fn is_over(&self) -> bool

Returns true if the mouse cursor is over the current entity.

pub fn lock_cursor_icon(&mut self)

Prevents the cursor icon from changing until the lock is released.

pub fn unlock_cursor_icon(&mut self)

Releases any cursor icon lock, allowing the cursor icon to be changed.

pub fn is_cursor_icon_locked(&self) -> bool

Returns true if the cursor icon is locked.

pub fn set_hover(&mut self, flag: bool)

Sets the hover flag of the current entity.

pub fn set_checked(&mut self, flag: bool)

Sets the checked flag of the current entity.

pub fn set_selected(&mut self, flag: bool)

Sets the checked flag of the current entity.

pub fn get_clipboard(
&mut self
) -> Result<String, Box<dyn Error + Sync + Send + 'static, Global>>

Get the contents of the system clipboard. This may fail for a variety of backend-specific reasons.

pub fn set_clipboard(
&mut self,
text: String
) -> Result<(), Box<dyn Error + Sync + Send + 'static, Global>>

Set the contents of the system clipboard. This may fail for a variety of backend-specific reasons.

pub fn toggle_class(&mut self, class_name: &str, applied: bool)

pub fn play_animation(&mut self, animation: Animation)

pub fn environment(&self) -> &Environment

pub fn needs_redraw(&mut self)

pub fn spawn<F>(&self, target: F) where
F: 'static + Send + FnOnce(&mut ContextProxy),

pub fn scale_factor(&self) -> f32