Struct vizia::prelude::List

pub struct List<L, T>where
    T: 'static,
    L: Lens<Target = Vec<T, Global>>,{ /* private fields */ }

A view for creating a list of items from a binding to a Vec

Implementations

impl<L, T> List<L, T>where
L: 'static + Lens<Target = Vec<T, Global>>,
T: Clone,

pub fn new<F>(cx: &mut Context, lens: L, item: F) -> Handle<'_, List<L, T>>where
F: 'static + Fn(&mut Context, usize, Then<L, Index<Vec<T, Global>, T>>),
<L as Lens>::Source: Model,

Creates a new List view with a binding to the given lens and a template for constructing the list items