vizia::prelude

Struct Runtime

pub struct Runtime { /* private fields */ }

The internal reactive Runtime which stores all the reactive system states in a thread local.

Implementations

impl Runtime

pub fn init_on_ui_thread()

Register the current thread as a UI thread.

Backends should call this when a UI event loop/window starts on a thread.

pub fn deinit_on_ui_thread()

Unregister the current thread as a UI thread.

Backends should call this when a UI event loop/window stops on a thread.

pub fn assert_ui_thread()

pub fn is_ui_thread() -> bool

pub fn drain_pending_work()

Drain any queued work from the sync runtime and run pending UI effects.

pub fn has_pending_work() -> bool

Returns true if there is queued work for this runtime or the sync runtime.

pub fn set_sync_effect_waker(waker: impl Fn() + Send + Sync + 'static)

Set a waker that will be called when a sync signal is updated off the UI thread. The waker should nudge the UI event loop (e.g., by sending a proxy event).

pub fn get_current_effect() -> Option<Rc<dyn EffectTrait>>

pub fn set_current_effect(effect: Option<Rc<dyn EffectTrait>>)

pub fn with_effect<T>( effect: Option<Rc<dyn EffectTrait>>, f: impl FnOnce() -> T, ) -> T