vizia::prelude

Trait AccessibilityModifiers

pub trait AccessibilityModifiers: Modifiable {
    // Provided methods
    fn role(self, role: Role) -> Self { ... }
    fn name<U>(self, name: impl Res<U>) -> Self
       where U: ToStringLocalized { ... }
    fn live(self, live: Live) -> Self { ... }
    fn hidden<U>(self, hidden: impl Res<U>) -> Self
       where U: Into<bool> { ... }
    fn numeric_value<U>(self, value: impl Res<U>) -> Self
       where U: Into<f64> { ... }
    fn text_value<U>(self, value: impl Res<U>) -> Self
       where U: ToStringLocalized { ... }
}

Modifiers for changing the accessibility properties of a view.

Provided Methods

fn role(self, role: Role) -> Self

Sets the accessibility role of the view.

fn name<U>(self, name: impl Res<U>) -> Self
where U: ToStringLocalized,

Sets the accessibility name of the view.

fn live(self, live: Live) -> Self

Sets whether the view should act as an accessibility live region.

fn hidden<U>(self, hidden: impl Res<U>) -> Self
where U: Into<bool>,

Sets whether the view should be hidden from accessibility.

fn numeric_value<U>(self, value: impl Res<U>) -> Self
where U: Into<f64>,

Sets the accessibility numeric value for the view.

fn text_value<U>(self, value: impl Res<U>) -> Self
where U: ToStringLocalized,

Sets the accessibility text value for the view.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.