Trait vizia::prelude::LensExt

pub trait LensExt: Lens {
    fn get<C>(&self, cx: &C) -> Self::Target
    where
        C: DataContext,
        Self::Target: Clone,
    { ... }
    fn get_fallible<C>(&self, cx: &C) -> Option<Self::Target>
    where
        C: DataContext,
        Self::Target: Clone,
    { ... }
    fn then<Other>(self, other: Other) -> Then<Self, Other>
    where
        Other: Lens<Source = Self::Target>,
    { ... }
    fn index<T>(self, index: usize) -> Then<Self, Index<Self::Target, T>>
    where
        T: 'static,
        Self::Target: Deref<Target = [T]>,
    { ... }
    fn map<G, B>(self, get: G) -> Then<Self, Map<G, Self::Target, B>>
    where
        G: Clone + 'static + Fn(&Self::Target) -> B,
        B: 'static + Clone,
    { ... }
    fn unwrap<T>(self) -> Then<Self, UnwrapLens<T>>
    where
        T: 'static,
        Self: Lens<Target = Option<T>>,
    { ... }
    fn into_lens<T>(self) -> Then<Self, IntoLens<Self::Target, T>>
    where
        T: 'static,
        Self::Target: Clone + Into<T>,
    { ... }
}

Helpers for constructing more complex Lenses.

This trait is part of the prelude.

Provided Methods

fn get<C>(&self, cx: &C) -> Self::Targetwhere
C: DataContext,
Self::Target: Clone,

Retrieve a copy of the lensed data from context.

Example

let value = lens.get(cx);

fn get_fallible<C>(&self, cx: &C) -> Option<Self::Target>where
C: DataContext,
Self::Target: Clone,

fn then<Other>(self, other: Other) -> Then<Self, Other>where
Other: Lens<Source = Self::Target>,

Used to construct a lens to some data contained within some other lensed data.

Example

Binds a label to other_data, which is a field of a struct SomeData, which is a field of the root AppData model:

Binding::new(cx, AppData::some_data.then(SomeData::other_data), |cx, data|{

});

fn index<T>(self, index: usize) -> Then<Self, Index<Self::Target, T>>where
T: 'static,
Self::Target: Deref<Target = [T]>,

fn map<G, B>(self, get: G) -> Then<Self, Map<G, Self::Target, B>>where
G: Clone + 'static + Fn(&Self::Target) -> B,
B: 'static + Clone,

fn unwrap<T>(self) -> Then<Self, UnwrapLens<T>>where
T: 'static,
Self: Lens<Target = Option<T>>,

fn into_lens<T>(self) -> Then<Self, IntoLens<Self::Target, T>>where
T: 'static,
Self::Target: Clone + Into<T>,