Trait vizia::prelude::Res

pub trait Res<T> {
    fn get_val(&self, cx: &Context) -> T;
    fn set_or_bind<F>(&self, cx: &mut Context, entity: Entity, closure: F)
    where
        F: 'static + Clone + Fn(&mut Context, Entity, T);

    fn get_val_fallible(&self, cx: &Context) -> Option<T> { ... }
}

A trait for types that can automatically resolve into other types, with or without consulting the Context.

This trait is part of the prelude.

Required Methods

fn get_val(&self, cx: &Context) -> T

fn set_or_bind<F>(&self, cx: &mut Context, entity: Entity, closure: F) where
F: 'static + Clone + Fn(&mut Context, Entity, T),

Provided Methods

fn get_val_fallible(&self, cx: &Context) -> Option<T>