vizia::prelude

Trait SliderModifiers

pub trait SliderModifiers: Sized {
    // Required methods
    fn on_change<F>(self, callback: F) -> Self
       where F: 'static + Fn(&mut EventContext<'_>, f32);
    fn range<U>(self, range: impl Res<U> + 'static) -> Self
       where U: Into<Range<f32>> + Clone + 'static;
    fn vertical<U>(self, vertical: impl Res<U> + 'static) -> Self
       where U: Into<bool> + Clone + 'static;
    fn step<U>(self, step: impl Res<U> + 'static) -> Self
       where U: Into<f32> + Clone + 'static;
    fn default_value<U>(self, default_value: impl Res<U> + 'static) -> Self
       where U: Into<f32> + Clone + 'static;
}

Required Methods

fn on_change<F>(self, callback: F) -> Self
where F: 'static + Fn(&mut EventContext<'_>, f32),

Sets the callback triggered when the slider value is changed.

Takes a closure which triggers when the slider value is changed, either by pressing the track or dragging the thumb along the track.

Slider::new(cx, value)
    .on_change(|cx, value| {
        let _ = (cx, value);
    });

fn range<U>(self, range: impl Res<U> + 'static) -> Self
where U: Into<Range<f32>> + Clone + 'static,

Sets the range of the slider.

If the source value is outside of the range then the slider will clip to min/max of the range.

Slider::new(cx, value)
    .range(-20.0..50.0)
    .on_change(|cx, value| {
        let _ = (cx, value);
    });

fn vertical<U>(self, vertical: impl Res<U> + 'static) -> Self
where U: Into<bool> + Clone + 'static,

Sets the orientation of the slider to vertical.

Slider::new(cx, value)
    .vertical(true)
    .on_change(|cx, value| {
        let _ = (cx, value);
    });

fn step<U>(self, step: impl Res<U> + 'static) -> Self
where U: Into<f32> + Clone + 'static,

Set the step value for the slider.

Slider::new(cx, value)
    .step(0.1_f32)
    .on_change(|cx, value| {
        let _ = (cx, value);
    });

fn default_value<U>(self, default_value: impl Res<U> + 'static) -> Self
where U: Into<f32> + Clone + 'static,

Sets the value that the slider resets to when the thumb is double-clicked.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.