Trait vizia::state::Model

pub trait Model: 'static + Sized {
    fn build(self, cx: &mut Context) { ... }
    fn event(&mut self, cx: &mut EventContext<'_>, event: &mut Event) { ... }
}

A trait implemented by application data in order to mutate in response to events.

Examples

pub struct AppData {
    count: i32,
}

enum AppEvent {
    Increment,
    Decrement,
}

impl Model for AppData {
    fn event(&mut self, cx: &mut EventContext, event: &mut Event) {
        event.map(|app_event, _| match app_event {
            AppEvent::Increment => {
                self.count += 1;
            }

            AppEvent::Decrement => {
                self.count -= 1;
            }
        });
    }
}

This trait is part of the prelude.

Provided Methods

fn build(self, cx: &mut Context)

Build the model data into the application tree.

Examples

fn main() {
    Application::new(|cx|{
        AppData::default().build(cx);
    }).run();  
}

fn event(&mut self, cx: &mut EventContext<'_>, event: &mut Event)

Respond to events in order to mutate the model data.

Examples

impl Model for AppData {
    fn event(&mut self, cx: &mut EventContext, event: &mut Event) {
        event.map(|app_event, _| match app_event {
            AppEvent::Increment => {
                self.count += 1;
            }

            AppEvent::Decrement => {
                self.count -= 1;
            }
        });
    }
}