Enum vizia::style::LengthValue

pub enum LengthValue {

    Px(f32),
    In(f32),
    Cm(f32),
    Mm(f32),
    Q(f32),
    Pt(f32),
    Pc(f32),
    Em(f32),
    Ex(f32),
    Ch(f32),
    Rem(f32),
    Vw(f32),
    Vh(f32),
    Vmin(f32),
    Vmax(f32),

}

A length value.

Variants

Px(f32)

Pixels.

A pixel is the smallest viewable unit. The length of a pixel depends on the used device.

In(f32)

Inches.

An inch is roughly 96 pixels or about 2.54cm.

Cm(f32)

Centimeters.

A centimeter is roughly 37.8 pixels.

Mm(f32)

Millimeters.

A millimeter is roughly 3.78 pixels or 1/10th of a centimeter.

Q(f32)

Quarter.

A quarter of a millimeter is roughly 1/40th of 1cm.

Pt(f32)

Points.

A point is roughly 1.3333 pixels or 1/72th of an inch.

Pc(f32)

Picas.

A pica is roughly 16 pixels or 1/6th of an inch.

Em(f32)

em.

An em is equal to the computed font-size property of the element on which it is used.

Ex(f32)

ex.

An ex is the height of the letter x of the font, or 0.5em.

Ch(f32)

Character unit.

A character unit is defined as the width of the character 0 of the font.

Rem(f32)

Root em.

An rem is equal to the computed font-size property of the root of the document. For websites the root of the document is the <html> element.

Vw(f32)

View width.

A length equal to 1% of the width of the viewport.

Vh(f32)

View height.

A length equal to 1% of the height of the viewport.

Vmin(f32)

Viewport minimum.

A viewport minimum is equal to 1% of the viewport’s smallest dimension. For example 1vmin is equal to 6px on a browser window that is 1200px wide and 600px high.

Vmax(f32)

Viewport maximum.

A viewport maximum is equal to 1% of the viewport’s largest dimension. For example 1vmax is equal to 12px on a browser window that is 1200px wide and 600px high.

Implementations

impl LengthValue

pub const PX_PER_IN: f32 = 96f32

The amount of pixels per inch.

pub const PX_PER_CM: f32 = 37.7952766f32

The amount of pixels per centimeter.

pub const PX_PER_MM: f32 = 3.77952766f32

The amount of pixels per millimeter.

pub const PX_PER_Q: f32 = 0.944881916f32

The amount of pixels per quarter-millimeter.

pub const PX_PER_PT: f32 = 1.33333337f32

The amount of pixels per point.

pub const PX_PER_PC: f32 = 16f32

The amount of pixels per pica.

pub fn to_px(&self) -> Option<f32>

Returns the amount of pixels of the length if calculateable.

pub fn to_unit_value(&self) -> (f32, &str)