Trait vizia::text::EditableText

pub trait EditableText: Clone {
    fn as_str(&self) -> &str;
    fn edit(&mut self, range: Range<usize>, txt: impl Into<String>);
    fn len(&self) -> usize;
    fn prev_grapheme_offset(&self, current: usize) -> Option<usize>;
    fn next_grapheme_offset(&self, current: usize) -> Option<usize>;
    fn prev_word_offset(&self, current: usize) -> Option<usize>;
    fn next_word_offset(&self, current: usize) -> Option<usize>;
    fn word_around(&self, pos: usize) -> RangeInclusive<usize>;
    fn paragraph_around(&self, pos: usize) -> RangeInclusive<usize>;
}

Required Methods

fn as_str(&self) -> &str

fn edit(&mut self, range: Range<usize>, txt: impl Into<String>)

Replace range with new text

fn len(&self) -> usize

Length of the text

fn prev_grapheme_offset(&self, current: usize) -> Option<usize>

Get the previous grapheme offset from the current offset if it exists

fn next_grapheme_offset(&self, current: usize) -> Option<usize>

Get the next grapheme offset from the current offset if it exists

fn prev_word_offset(&self, current: usize) -> Option<usize>

Get the prev word offset from the current offset if it exists

fn next_word_offset(&self, current: usize) -> Option<usize>

fn word_around(&self, pos: usize) -> RangeInclusive<usize>

fn paragraph_around(&self, pos: usize) -> RangeInclusive<usize>