Module vizia::tree

Tree of Widgets

The Tree struct describes the visual hierarchy of widgets built into the application. A series of iterators are used to traverse the tree, which is used by nearly all systems, including for example, for calculating layout, propagating events, and rendering the UI.

Structs

ChildIterator

Iterator for iterating the children of an entity.

DoubleEndedTreeTour

Double ended version of TreeTour.

ParentIterator

Iterator for iterating through the ancestors of an entity.

Tree

The Tree describes the tree of entities.

TreeDepthIterator

TreeIterator

Iterator for iterating through the tree in depth first preorder.

TreeTour

Modular tree traversal helper. Based on the Euler tour technique.

TreeTourIterator

Generic iterator based on TreeTour.

Enums

TourDirection

Current traversal direction.

TourStep

Where to go next after entering or leaving a node.

TreeError

Traits

TreeExt

Trait which provides methods for querying the tree.

Functions

focus_backward

focus_forward

is_navigatable