Struct vizia::vg::canvas::OwnedCanvas

pub struct OwnedCanvas<'lt>(/* private fields */);

Represents a Canvas that is owned and dropped when it goes out of scope and is bound to the lifetime of some other value (an array of pixels for example).

Access to the Canvas functions are resolved with the Deref trait.