vizia::vg

Module image

Modules

images

Structs

CubicResampler

Specify b and c (each between 0…1) to create a shader that applies the corresponding cubic reconstruction filter to the image.

RequiredProperties

Enums

BitDepth

CachingHint

CompressionType

Type Aliases

Image

Image describes a two dimensional array of pixels to draw. The pixels may be decoded in a raster bitmap, encoded in a Picture or compressed data stream, or located in GPU memory as a GPU texture.