vizia::vg::image_filters

Function drop_shadow_only

pub fn drop_shadow_only(
    offset: impl Into<Point>,
    _: (f32, f32),
    color: impl Into<Color4f>,
    color_space: impl Into<Option<RCHandle<SkColorSpace>>>,
    input: impl Into<Option<RCHandle<SkImageFilter>>>,
    crop_rect: impl Into<CropRect>,
) -> Option<RCHandle<SkImageFilter>>

Create a filter that renders a drop shadow, in exactly the same manner as ::DropShadow, except that the resulting image does not include the input content. This allows the shadow and input to be composed by a filter DAG in a more flexible manner.

offset - The offset of the shadow.

sigma_x - The blur radius for the shadow, along the X axis.

sigma_y - The blur radius for the shadow, along the Y axis.

color - The color of the drop shadow.

color_space - The color space of the drop shadow color.

input - The input filter, or will use the source bitmap if this is null.

crop_rect - Optional rectangle that crops the input and output.