Trait vizia::vg::imgref::ImgExtMut

pub trait ImgExtMut<Pixel> {
    fn rows_padded_mut(&mut self) -> ChunksMut<'_, Pixel>;
    fn as_mut(&mut self) -> Img<&mut [Pixel]>;
}

Additional methods that depend on buffer size

To use these methods you need:

use imgref::*;

Required Methods

fn rows_padded_mut(&mut self) -> ChunksMut<'_, Pixel>

Iterate over the entire buffer as rows, including all padding

Rows will have up to stride width, but the last row may be shorter.

fn as_mut(&mut self) -> Img<&mut [Pixel]>

Borrow the container mutably