Type Alias vizia::vg::imgref::ImgRefMut

pub type ImgRefMut<'a, Pixel> = Img<&'a mut [Pixel]>;

Same as ImgRef, but mutable Pass this structure by value (i.e. ImgRef, not &ImgRef).

Aliased Type

struct ImgRefMut<'a, Pixel> {
    pub buf: &'a mut [Pixel],
    pub stride: usize,
    pub width: u32,
    pub height: u32,
}

Fields
buf: &'a mut [Pixel]
👎Deprecated: Don’t access struct fields directly. Use buf(), buf_mut() or into_buf()

Storage for the pixels. Usually Vec<Pixel> or &[Pixel]. See ImgVec and ImgRef.

Note that future version will make this field private. Use .rows() and .pixels() iterators where possible, or buf()/buf_mut()/into_buf().

stride: usize
👎Deprecated: Don’t access struct fields directly. Use stride()

Number of pixels to skip in the container to advance to the next row.

Note: pixels between width and stride may not be usable, and may not even exist in the last row.

width: u32
👎Deprecated: Don’t access struct fields directly. Use width()

Width of the image in pixels.

Note that this isn’t same as the width of the row in the buf, see stride

height: u32
👎Deprecated: Don’t access struct fields directly. Use height()

Height of the image in pixels.

Implementations

impl<Container> Img<Container>

pub fn width(&self) -> usize

Width of the image in pixels.

Note that this isn’t same as the width of the row in image data, see stride()

pub fn height(&self) -> usize

Height of the image in pixels.

pub fn stride(&self) -> usize

Number of pixels to skip in the container to advance to the next row.

Note the last row may have fewer pixels than the stride. Some APIs use number of bytes for a stride. You may need to multiply this one by number of pixels.

pub fn buf(&self) -> &Container

Immutable reference to the pixel storage. Warning: exposes stride. Use pixels() or rows() instead.

See also into_contiguous_buf().

pub fn buf_mut(&mut self) -> &mut Container

Mutable reference to the pixel storage. Warning: exposes stride. Use pixels_mut() or rows_mut() instead.

See also into_contiguous_buf().

pub fn into_buf(self) -> Container

Get the pixel storage by consuming the image. Be careful about stride — see into_contiguous_buf() for a safe version.

impl<'a, T> Img<&'a mut [T]>

pub fn sub_image( &'a mut self, left: usize, top: usize, width: usize, height: usize ) -> Img<&'a [T]>

Turn this into immutable reference, and slice a subregion of it

pub fn sub_image_mut( &mut self, left: usize, top: usize, width: usize, height: usize ) -> Img<&mut [T]>

Trim this image without copying. Note that mutable borrows are exclusive, so it’s not possible to have more than one mutable subimage at a time.

pub fn as_ref(&self) -> Img<&[T]>

Make mutable reference immutable

impl<'a, T> Img<&'a mut [T]>where T: Copy,

pub fn pixels(&self) -> PixelsIter<'_, T>

Panics

If you want to iterate in parallel, parallelize rows() instead.

if width is 0

pub fn pixels_mut(&mut self) -> PixelsIterMut<'_, T>

If you want to iterate in parallel, parallelize rows() instead.

Panics

if width is 0

impl<'a, T> Img<&'a mut [T]>

pub fn rows(&self) -> RowsIter<'_, T>

Iterate over whole rows as slices

Panics

if stride is 0

pub fn rows_mut(&mut self) -> RowsIterMut<'_, T>

Iterate over whole rows as slices

Panics

if stride is 0

impl<Container> Img<Container>

pub fn new_stride( buf: Container, width: usize, height: usize, stride: usize ) -> Img<Container>

Same as new(), except each row is located stride number of pixels after the previous one.

Stride can be equal to width or larger. If it’s larger, then pixels between end of previous row and start of the next are considered a padding, and may be ignored.

The Container is usually a Vec or a slice.

pub fn new(buf: Container, width: usize, height: usize) -> Img<Container>

Create new image with Container (which can be Vec, &[] or something else) with given width and height in pixels.

Assumes the pixels in container are contiguous, layed out row by row with width pixels per row and at least height rows.

If the container is larger than width×height pixels, the extra rows are a considered a padding and may be ignored.

impl<OldContainer> Img<OldContainer>

pub fn map_buf<NewContainer, OldPixel, NewPixel, F>( self, callback: F ) -> Img<NewContainer>where NewContainer: AsRef<[NewPixel]>, OldContainer: AsRef<[OldPixel]>, F: FnOnce(OldContainer) -> NewContainer,

A convenience method for creating an image of the same size and stride, but with a new buffer.

pub fn new_buf<NewContainer, OldPixel, NewPixel>( &self, new_buf: NewContainer ) -> Img<NewContainer>where NewContainer: AsRef<[NewPixel]>, OldContainer: AsRef<[OldPixel]>,

A convenience method for creating an image of the same size and stride, but with a new buffer.

impl<T> Img<T>where T: ToOwned,

pub fn to_owned(&self) -> Img<<T as ToOwned>::Owned>

Convert underlying buffer to owned (e.g. slice to vec)

See also to_contiguous_buf().0.into_owned()