Type Definition vizia::vg::imgref::ImgVec

pub type ImgVec<Pixel> = Img<Vec<Pixel, Global>>;

Image owning its pixels.

A 2D array of pixels. The pixels are oriented top-left first and rows are stride pixels wide.

If size of the buf is larger than width*height, then any excess space is a padding (see width_padded()/height_padded()).