Struct vizia::vg::pdf::Metadata

pub struct Metadata {

    pub title: String,
    pub author: String,
    pub subject: String,
    pub keywords: String,
    pub creator: String,
    pub producer: String,
    pub creation: Option<DateTime>,
    pub modified: Option<DateTime>,
    pub lang: String,
    pub raster_dpi: Option<f32>,
    pub pdf_a: bool,
    pub encoding_quality: Option<i32>,
    pub structure_element_tree_root: Option<StructureElementNode>,
    pub compression_level: SkPDF_Metadata_CompressionLevel,

}

Optional metadata to be passed into the PDF factory function.

Fields
title: String

The document’s title.

author: String

The name of the person who created the document.

subject: String

The subject of the document.

keywords: String

Keywords associated with the document. Commas may be used to delineate keywords within the string.

creator: String

If the document was converted to PDF from another format, the name of the conforming product that created the original document from which it was converted.

producer: String

The product that is converting this document to PDF.

creation: Option<DateTime>

The date and time the document was created.

modified: Option<DateTime>

The date and time the document was most recently modified.

lang: String

The natural language of the text in the PDF. If lang is empty, the root StructureElementNode::lang will be used (if not empty). Text not in this language should be marked with StructureElementNode::lang.

raster_dpi: Option<f32>

The DPI (pixels-per-inch) at which features without native PDF support will be rasterized (e.g. draw image with perspective, draw text with perspective, …) A larger DPI would create a PDF that reflects the original intent with better fidelity, but it can make for larger PDF files too, which would use more memory while rendering, and it would be slower to be processed or sent online or to printer.

pdf_a: bool

If true, include XMP metadata, a document UUID, and s_rgb output intent information. This adds length to the document and makes it non-reproducible, but are necessary features for PDF/A-2b conformance

encoding_quality: Option<i32>

Encoding quality controls the trade-off between size and quality. By default this is set to 101 percent, which corresponds to lossless encoding. If this value is set to a value <= 100, and the image is opaque, it will be encoded (using JPEG) with that quality setting.

structure_element_tree_root: Option<StructureElementNode>compression_level: SkPDF_Metadata_CompressionLevel

PDF streams may be compressed to save space. Use this to specify the desired compression vs time tradeoff.