Struct vizia::vg::renderer::OpenGl

pub struct OpenGl { /* private fields */ }

Implementations

impl OpenGl

pub unsafe fn new_from_function<F>(load_fn: F) -> Result<OpenGl, ErrorKind> where
F: FnMut(&str) -> *const c_void,

pub fn new_from_glutin_context(
windowed_context: &ContextWrapper<PossiblyCurrent, Window>
) -> Result<OpenGl, ErrorKind>

pub fn is_opengles(&self) -> bool

pub fn set_screen_target(
&mut self,
framebuffer_object: Option<<Context as HasContext>::Framebuffer>
)

Make the “Screen” RenderTarget actually render to a framebuffer object. This is useful when embedding femtovg into another program where final composition is handled by an external task. The given framebuffer_object must refer to a Framebuffer Object created on the current OpenGL Context, and must have a depth & stencil attachment.

Pass None to clear any previous Framebuffer Object ID that was passed and target rendering to the default target (normally the window).