Trait vizia::vg::rgb::AsPixels

pub trait AsPixels<PixelType> {
    fn as_pixels(&self) -> &[PixelType] ;
    fn as_pixels_mut(&mut self) -> &mut [PixelType] ;
}

Casts a slice of bytes into a slice of pixels, e.g. [u8] to [RGB8].

See also FromSlice

Required Methods

fn as_pixels(&self) -> &[PixelType]

Reinterpret the slice as a read-only/shared slice of pixels. Multiple consecutive elements in the slice are intepreted as a single pixel (depending on format, e.g. 3 for RGB, 4 for RGBA).

Leftover elements are ignored if the slice isn’t evenly divisible into pixels.

Use this method only when the type is known from context. See also FromSlice.

fn as_pixels_mut(&mut self) -> &mut [PixelType]

Reinterpret the slice as a mutable/exclusive slice of pixels. Multiple consecutive elements in the slice are intepreted as a single pixel (depending on format, e.g. 3 for RGB, 4 for RGBA).

Leftover elements are ignored if the slice isn’t evenly divisible into pixels.

Use this method only when the type is known from context. See also FromSlice.