vizia::vg::yuva_info
Type Alias YUVAInfo

pub type YUVAInfo = Handle<SkYUVAInfo>;

Specifies the structure of planes for a YUV image with optional alpha. The actual planar data is not part of this structure and depending on usage is in external textures or pixmaps.

Aliased Type

struct YUVAInfo(/* private fields */);