vizia::view

Struct Handle

pub struct Handle<'a, V> { /* private fields */ }

A handle to a view which has been built into the tree.

Implementations

impl<V> Handle<'_, V>

pub fn entity(&self) -> Entity

Returns the [Entity] id of the view.

pub fn context(&mut self) -> &mut Context

Returns a mutable reference to the context.

pub fn parent(&self) -> Entity

Returns the entity id of the parent view.

pub fn lock_focus_to_within(self) -> Handle<'_, V>

Stop the user from tabbing out of a subtree, which is useful for modal dialogs.

pub fn modify<F>(self, f: F) -> Handle<'_, V>
where F: FnOnce(&mut V), V: 'static,

Mody the internal data of the view.

pub fn on_build<F>(self, callback: F) -> Handle<'_, V>
where F: Fn(&mut EventContext<'_>),

Callback which is run when the view is built/rebuilt.

pub fn bind<R, T, F>(self, signal: R, closure: F) -> Handle<'_, V>
where R: SignalGet<T> + 'static, T: Clone + 'static, F: 'static + Fn(Handle<'_, V>),

Creates a binding to the given resource and provides a closure which can be used to mutate the view through a handle.

pub fn needs_relayout(&mut self)

Marks the view as needing a relayout.

pub fn needs_restyle(&mut self)

Marks the view as needing a restyle.

pub fn needs_redraw(&mut self)

Marks the view as needing a redraw.

pub fn bounds(&self) -> BoundingBox

Returns the bounding box of the view.

pub fn scale_factor(&self) -> f32

Returns the scale factor of the device.

impl Handle<'_, Avatar>

pub fn variant<U>(self, variant: impl Res<U> + 'static) -> Handle<'_, Avatar>
where U: Into<AvatarVariant> + Clone + PartialEq + 'static,

Selects the geometric variant of the Avatar. Accepts a value or signal of type AvatarVariant.

Avatar::new(cx, |cx|{
    Svg::new(cx, ICON_USER);
})
.variant(AvatarVariant::Rounded);

pub fn badge<F>(self, content: F) -> Handle<'_, Avatar>
where F: FnOnce(&mut Context) -> Handle<'_, Badge>,

Adds a badge to the Avatar.

Avatar::new(cx, |cx|{
    Svg::new(cx, ICON_USER);
})
.badge(|cx| Badge::empty(cx).class("error"));

impl Handle<'_, Badge>

pub fn placement<U>(self, placement: impl Res<U> + 'static) -> Handle<'_, Badge>
where U: Into<BadgePlacement> + Clone + 'static,

Sets the placement of a badge relative to its parent. Accepts a value or signal of type BadgePlacement.

impl Handle<'_, ButtonGroup>

pub fn vertical(self, is_vertical: impl Res<bool>) -> Handle<'_, ButtonGroup>

Sets whether the button group is in vertical orientation.

impl Handle<'_, Checkbox>

pub fn on_toggle<F>(self, callback: F) -> Handle<'_, Checkbox>
where F: 'static + Fn(&mut EventContext<'_>),

Set the callback triggered when the checkbox is pressed.

Examples

Checkbox::new(cx, value)
    .on_toggle(|cx| cx.emit(AppEvent::ToggleValue));

impl Handle<'_, Chip>

pub fn on_close( self, callback: impl Fn(&mut EventContext<'_>) + Send + Sync + 'static, ) -> Handle<'_, Chip>

Set the callback triggered when the close button of the chip is pressed. The chip close button is not displayed by default. Setting this callback causes the close button to be displayed.

pub fn variant<U>(self, variant: impl Res<U> + 'static) -> Handle<'_, Chip>
where U: Into<ChipVariant> + Clone + 'static,

Selects the style variant to be used by the chip. Accepts a value or signal of type ChipVariant.

Example

Chip::new(cx, "Chip")
    .variant(ChipVariant::Filled);

impl Handle<'_, Collapsible>

pub fn open(self, open: impl Res<bool> + 'static) -> Handle<'_, Collapsible>

Set the open state of the collapsible view.

impl<L, S, T> Handle<'_, ComboBox<L, S, T>>
where L: SignalGet<Vec<T>> + 'static, T: 'static + Clone + ToString, S: SignalGet<usize> + 'static,

pub fn on_select<F>(self, callback: F) -> Handle<'_, ComboBox<L, S, T>>
where F: 'static + Fn(&mut EventContext<'_>, usize),

Sets the callback triggered when an item is selected from the ComboBox list.

impl Handle<'_, Datepicker>

pub fn on_select<F>(self, callback: F) -> Handle<'_, Datepicker>
where F: 'static + Fn(&mut EventContext<'_>, NaiveDate),

Set the callback triggered when a date is selected from the Datepicker view.

impl Handle<'_, Divider>

pub fn orientation( self, orientation: impl Res<Orientation> + 'static, ) -> Handle<'_, Divider>

Set the orientation of the divider. Accepts a value or signal of type Orientation.

impl Handle<'_, Dropdown>

pub fn placement( self, placement: impl Res<Placement> + 'static, ) -> Handle<'_, Dropdown>

Sets the position where the tooltip should appear relative to its parent element. Defaults to Placement::Bottom.

pub fn show_arrow( self, show_arrow: impl Res<bool> + 'static, ) -> Handle<'_, Dropdown>

Sets whether the popup should include an arrow. Defaults to true.

pub fn arrow_size<U>(self, size: impl Res<U> + 'static) -> Handle<'_, Dropdown>
where U: Into<Length> + Clone + 'static,

Sets the size of the popup arrow, or gap if the arrow is hidden.

pub fn should_reposition( self, flag: impl Res<bool> + 'static, ) -> Handle<'_, Dropdown>

Set to whether the popup should reposition to always be visible.

impl<T> Handle<'_, Knob<T>>
where T: Res<f32> + 'static,

pub fn on_change<F>(self, callback: F) -> Handle<'_, Knob<T>>
where F: 'static + Fn(&mut EventContext<'_>, f32),

Sets the callback triggered when the knob value is changed.

impl Handle<'_, ArcTrack>

pub fn value<R>(self, value: R) -> Handle<'_, ArcTrack>
where R: Res<f32>,

impl Handle<'_, TickKnob>

pub fn value<R>(self, value: R) -> Handle<'_, TickKnob>
where R: Res<f32>,

impl Handle<'_, Label>

pub fn describing( self, entity_identifier: impl Into<String>, ) -> Handle<'_, Label>

Which form element does this label describe.

Examples

Checkbox::new(cx, value).on_toggle(|cx| cx.emit(AppEvent::ToggleValue)).id("checkbox_identifier");
Label::new(cx, "hello").describing("checkbox_identifier");

impl Handle<'_, List>

pub fn selected<R>(self, selected: impl Res<R> + 'static) -> Handle<'_, List>
where R: Deref<Target = [usize]> + Clone + 'static,

Sets the selected items of the list from signal of type indices.

pub fn on_select<F>(self, callback: F) -> Handle<'_, List>
where F: 'static + Fn(&mut EventContext<'_>, usize),

Sets the callback triggered when a ListItem is selected.

pub fn selectable<U>( self, selectable: impl Res<U> + 'static, ) -> Handle<'_, List>
where U: Into<Selectable> + Clone + 'static,

Set the selectable state of the List.

pub fn selection_follows_focus<U>( self, flag: impl Res<U> + 'static, ) -> Handle<'_, List>
where U: Into<bool> + Clone + 'static,

Sets whether the selection should follow the focus.

pub fn orientation<U>( self, orientation: impl Res<U> + 'static, ) -> Handle<'_, List>
where U: Into<Orientation> + Clone + 'static,

Sets the orientation of the list.

pub fn scroll_to_cursor(self, flag: bool) -> Handle<'_, List>

Sets whether the scrollbar should move to the cursor when pressed.

pub fn on_scroll( self, callback: impl Fn(&mut EventContext<'_>, f32, f32) + Send + Sync + 'static, ) -> Handle<'_, List>

Sets a callback which will be called when a scrollview is scrolled, either with the mouse wheel, touchpad, or using the scroll bars.

pub fn scroll_x(self, scrollx: impl Res<f32> + 'static) -> Handle<'_, List>

Set the horizontal scroll position of the ScrollView. Accepts a value or signal of type an f32 between 0 and 1.

pub fn scroll_y(self, scrollx: impl Res<f32> + 'static) -> Handle<'_, List>

Set the vertical scroll position of the ScrollView. Accepts a value or signal of type an f32 between 0 and 1.

pub fn show_horizontal_scrollbar( self, flag: impl Res<bool> + 'static, ) -> Handle<'_, List>

Sets whether the horizontal scrollbar should be visible.

pub fn show_vertical_scrollbar( self, flag: impl Res<bool> + 'static, ) -> Handle<'_, List>

Sets whether the vertical scrollbar should be visible.

impl Handle<'_, PickList>

pub fn placeholder<P>( self, placeholder: impl Res<P> + 'static, ) -> Handle<'_, PickList>
where P: ToStringLocalized + Clone + 'static,

Sets the placeholder text that appears when the textbox has no value.

pub fn on_select<F>(self, callback: F) -> Handle<'_, PickList>
where F: 'static + Fn(&mut EventContext<'_>, usize),

Sets the callback triggered when an option is selected.

impl Handle<'_, Popup>

pub fn placement( self, placement: impl Res<Placement> + 'static, ) -> Handle<'_, Popup>

Sets the position where the popup should appear relative to its parent element. Defaults to Placement::Bottom.

pub fn show_arrow( self, show_arrow: impl Res<bool> + 'static, ) -> Handle<'_, Popup>

Sets whether the popup should include an arrow. Defaults to true.

pub fn arrow_size<U>(self, size: impl Res<U> + 'static) -> Handle<'_, Popup>
where U: Into<Length> + Clone + 'static,

Sets the size of the popup arrow, or gap if the arrow is hidden.

pub fn should_reposition( self, should_reposition: impl Res<bool> + 'static, ) -> Handle<'_, Popup>

Set to whether the popup should reposition to always be visible.

pub fn on_blur<F>(self, f: F) -> Handle<'_, Popup>
where F: 'static + Fn(&mut EventContext<'_>),

Registers a callback for when the user clicks off of the popup, usually with the intent of closing it.

impl Handle<'_, RadioButton>

pub fn on_select<F>(self, callback: F) -> Handle<'_, RadioButton>
where F: 'static + Fn(&mut EventContext<'_>),

Set the callback triggered when the radio button is selected.

Examples

RadioButton::new(cx, value)
    .on_select(|cx| cx.emit(AppEvent::ToggleValue));

impl<R> Handle<'_, Rating<R>>
where R: SignalGet<u32> + SignalUpdate<u32> + SignalMap<u32> + Res<u32> + 'static,

pub fn on_change<F>(self, callback: F) -> Handle<'_, Rating<R>>
where F: 'static + Fn(&mut EventContext<'_>, u32),

Set the callback which is triggered when the rating changes.

impl Handle<'_, ResizableStack>

pub fn on_reset<F>(self, on_reset: F) -> Handle<'_, ResizableStack>
where F: Fn(&mut EventContext<'_>) + 'static,

Sets a callback to be called when the stack is reset, i.e. when the resize handle is double-clicked.

impl Handle<'_, Scrollbar>

pub fn scroll_to_cursor( self, scroll_to_cursor: impl Res<bool> + 'static, ) -> Handle<'_, Scrollbar>

Sets whether the scrollbar should move to the cursor when pressed.

impl Handle<'_, ScrollView>

pub fn on_scroll( self, callback: impl Fn(&mut EventContext<'_>, f32, f32) + Send + Sync + 'static, ) -> Handle<'_, ScrollView>

Sets a callback which will be called when a scrollview is scrolled, either with the mouse wheel, touchpad, or using the scroll bars.

pub fn scroll_to_cursor( self, scroll_to_cursor: impl Res<bool> + 'static, ) -> Handle<'_, ScrollView>

Sets whether the scrollbar should move to the cursor when pressed.

pub fn scroll_x( self, scrollx: impl Res<f32> + 'static, ) -> Handle<'_, ScrollView>

Set the horizontal scroll position of the ScrollView. Accepts a value or signal of type f32 between 0 and 1.

pub fn scroll_y( self, scrolly: impl Res<f32> + 'static, ) -> Handle<'_, ScrollView>

Set the vertical scroll position of the ScrollView. Accepts a value or signal of type f32 between 0 and 1.

pub fn show_horizontal_scrollbar( self, flag: impl Res<bool> + 'static, ) -> Handle<'_, ScrollView>

Sets whether the horizontal scrollbar should be visible.

pub fn show_vertical_scrollbar( self, flag: impl Res<bool> + 'static, ) -> Handle<'_, ScrollView>

Sets whether the vertical scrollbar should be visible.

impl<S> Handle<'_, Slider<S>>
where S: SignalGet<f32> + 'static,

pub fn on_change<F>(self, callback: F) -> Handle<'_, Slider<S>>
where F: 'static + Fn(&mut EventContext<'_>, f32),

Sets the callback triggered when the slider value is changed.

Takes a closure which triggers when the slider value is changed, either by pressing the track or dragging the thumb along the track.

Slider::new(cx, value)
    .on_change(|cx, value| {
        let _ = (cx, value);
    });

pub fn range<U>(self, range: impl Res<U> + 'static) -> Handle<'_, Slider<S>>
where U: Into<Range<f32>> + Clone + 'static,

Sets the range of the slider.

If the source value is outside of the range then the slider will clip to min/max of the range.

Slider::new(cx, value)
    .range(-20.0..50.0)
    .on_change(|cx, value| {
        let _ = (cx, value);
    });

pub fn orientation<U>( self, orientation: impl Res<U> + 'static, ) -> Handle<'_, Slider<S>>
where U: Into<Orientation> + Clone + 'static,

Sets the orientation of the slider.

Slider::new(cx, value)
    .orientation(Orientation::Vertical)
    .on_change(|cx, value| {
        let _ = (cx, value);
    });

pub fn step<U>(self, step: impl Res<U> + 'static) -> Handle<'_, Slider<S>>
where U: Into<f32> + Clone + 'static,

Set the step value for the slider.

Slider::new(cx, value)
    .step(0.1_f32)
    .on_change(|cx, value| {
        let _ = (cx, value);
    });

pub fn keyboard_fraction<U>( self, keyboard_fraction: impl Res<U> + 'static, ) -> Handle<'_, Slider<S>>
where U: Into<f32> + Clone + 'static,

Sets the fraction of a slider that a press of an arrow key will change.

Slider::new(cx, value)
    .keyboard_fraction(0.05_f32)
    .on_change(|cx, value| {
        let _ = (cx, value);
    });

impl Handle<'_, Spinbox>

pub fn on_increment<F>(self, callback: F) -> Handle<'_, Spinbox>
where F: 'static + Fn(&mut EventContext<'_>) + Send + Sync,

Sets the callback which is triggered when the Spinbox value is incremented.

pub fn on_decrement<F>(self, callback: F) -> Handle<'_, Spinbox>
where F: 'static + Fn(&mut EventContext<'_>) + Send + Sync,

Sets the callback which is triggered when the Spinbox value is decremented.

pub fn orientation( self, orientation: impl Res<Orientation> + 'static, ) -> Handle<'_, Spinbox>

Sets the orientation of the Spinbox.

pub fn icons( self, icons: impl Res<SpinboxIcons> + 'static, ) -> Handle<'_, Spinbox>

Set the icons which should be used for the increment and decrement buttons of the Spinbox

impl Handle<'_, Switch>

pub fn on_toggle<F>(self, callback: F) -> Handle<'_, Switch>
where F: 'static + Fn(&mut EventContext<'_>),

Set the callback triggered when the Switch is pressed.

Examples

Switch::new(cx, value)
    .on_toggle(|cx| cx.emit(AppEvent::ToggleValue));

impl Handle<'_, TabView>

pub fn vertical(self) -> Handle<'_, TabView>

pub fn on_select( self, callback: impl Fn(&mut EventContext<'_>, usize) + 'static, ) -> Handle<'_, TabView>

pub fn with_selected<U>(self, selected: impl Res<U>) -> Handle<'_, TabView>
where U: Into<usize>,

impl<R, T> Handle<'_, Textbox<R, T>>
where R: Res<T> + 'static, T: Clone + ToStringLocalized + FromStr + 'static,

pub fn on_edit<F>(self, callback: F) -> Handle<'_, Textbox<R, T>>
where F: 'static + Fn(&mut EventContext<'_>, String) + Send + Sync,

Sets the callback triggered when a textbox is edited, i.e. text is inserted/deleted.

Callback provides the current text of the textbox.

pub fn on_submit<F>(self, callback: F) -> Handle<'_, Textbox<R, T>>
where F: 'static + Fn(&mut EventContext<'_>, T, bool) + Send + Sync,

Sets the callback triggered when a textbox is submitted, i.e. when the enter key is pressed with a single-line textbox or the textbox loses focus.

Callback provides the text of the textbox and a flag to indicate if the submit was due to a key press or a loss of focus.

pub fn on_blur<F>(self, callback: F) -> Handle<'_, Textbox<R, T>>
where F: 'static + Fn(&mut EventContext<'_>) + Send + Sync,

Sets the callback triggered when a textbox is blurred, i.e. the mouse is pressed outside of the textbox.

pub fn on_cancel<F>(self, callback: F) -> Handle<'_, Textbox<R, T>>
where F: 'static + Fn(&mut EventContext<'_>) + Send + Sync,

Sets the callback triggered when a textbox edit is cancelled, i.e. the escape key is pressed while editing.

pub fn validate<F>(self, is_valid: F) -> Handle<'_, Textbox<R, T>>
where F: 'static + Fn(&T) -> bool + Send + Sync,

Sets a validation closure which is called when the textbox is edited and sets the validity attribute to the output of the closure.

If a textbox is modified with the validate modifier then the on_submit will not be called if the text is invalid.

pub fn placeholder<P>( self, text: impl Res<P> + 'static, ) -> Handle<'_, Textbox<R, T>>
where P: ToStringLocalized + Clone + 'static,

Sets the placeholder text that appears when the textbox has no value.

impl Handle<'_, ToggleButton>

pub fn on_toggle( self, callback: impl Fn(&mut EventContext<'_>) + 'static, ) -> Handle<'_, ToggleButton>

Sets the callback triggered when the ToggleButton is toggled.

impl Handle<'_, Tooltip>

pub fn placement<U>( self, placement: impl Res<U> + 'static, ) -> Handle<'_, Tooltip>
where U: Into<Placement> + Clone + 'static,

Sets the position where the tooltip should appear relative to its parent element. Defaults to Placement::Bottom.

pub fn arrow<U>(self, show_arrow: impl Res<U> + 'static) -> Handle<'_, Tooltip>
where U: Into<bool> + Clone + 'static,

Sets whether the tooltip should include an arrow. Defaults to true.

pub fn arrow_size<U>(self, size: impl Res<U> + 'static) -> Handle<'_, Tooltip>
where U: Into<Length> + Clone + 'static,

Sets the size of the tooltip arrow if enabled.

impl Handle<'_, VirtualList>

pub fn selected<R>( self, selected: impl Res<R> + 'static, ) -> Handle<'_, VirtualList>
where R: Deref<Target = [usize]> + Clone + 'static,

Sets the selected items of the list from a signal of indices.

pub fn on_select<F>(self, callback: F) -> Handle<'_, VirtualList>
where F: 'static + Fn(&mut EventContext<'_>, usize),

Sets the callback triggered when a ListItem is selected.

pub fn selectable<U>( self, selectable: impl Res<U> + 'static, ) -> Handle<'_, VirtualList>
where U: Into<Selectable> + Clone + 'static,

Set the selectable state of the List.

pub fn selection_follows_focus<U>( self, flag: impl Res<U> + 'static, ) -> Handle<'_, VirtualList>
where U: Into<bool> + Clone + 'static,

Sets whether the selection should follow the focus.

pub fn scroll_to_cursor(self, flag: bool) -> Handle<'_, VirtualList>

Sets whether the scrollbar should move to the cursor when pressed.

pub fn on_scroll( self, callback: impl Fn(&mut EventContext<'_>, f32, f32) + Send + Sync + 'static, ) -> Handle<'_, VirtualList>

Sets a callback which will be called when a scrollview is scrolled, either with the mouse wheel, touchpad, or using the scroll bars.

pub fn scroll_x( self, scrollx: impl Res<f32> + 'static, ) -> Handle<'_, VirtualList>

Set the horizontal scroll position of the ScrollView. Accepts a value or lens to an ‘f32’ between 0 and 1.

pub fn scroll_y( self, scrollx: impl Res<f32> + 'static, ) -> Handle<'_, VirtualList>

Set the vertical scroll position of the ScrollView. Accepts a value or lens to an ‘f32’ between 0 and 1.

pub fn show_horizontal_scrollbar( self, flag: impl Res<bool> + 'static, ) -> Handle<'_, VirtualList>

Sets whether the horizontal scrollbar should be visible.

pub fn show_vertical_scrollbar( self, flag: impl Res<bool> + 'static, ) -> Handle<'_, VirtualList>

Sets whether the vertical scrollbar should be visible.

impl Handle<'_, XYPad>

pub fn on_change<F>(self, callback: F) -> Handle<'_, XYPad>
where F: Fn(&mut EventContext<'_>, f32, f32) + 'static,

Set the callback which will be triggered when the XYPad is manipulated.