vizia::views

Enum MenuEvent

pub enum MenuEvent {

    ToggleOpen,
    Open,
    Close,
    CloseAndFocusTrigger,
    CloseAll,
    MenuIsOpen,
    FocusNext,
    FocusPrev,
    FocusFirst,
    FocusLast,
    TriggerArrowDown,
    TriggerArrowRight,
    TriggerArrowLeft,
    FocusNextMenuBarItem,
    FocusPrevMenuBarItem,

}

Events used by menus.

Variants

ToggleOpen

Toggle the open state of the menu.

Open

Sets the menu to an open state.

Close

Sets the menu to a closed state.

CloseAndFocusTrigger

Closes the active menu and restores focus to the trigger that opened it.

CloseAll

Closes the menu and any submenus.

MenuIsOpen

Event emitted when a menu or submenu is opened.

FocusNext

Move focus to the next item within the open popup.

FocusPrev

Move focus to the previous item within the open popup.

FocusFirst

Move focus to the first item within the open popup.

FocusLast

Move focus to the last item within the open popup.

TriggerArrowDown

ArrowDown pressed on a submenu trigger (open popup or within popup).

TriggerArrowRight

ArrowRight pressed on a submenu trigger (navigate or open submenu).

TriggerArrowLeft

ArrowLeft pressed on a submenu trigger (navigate or close submenu).

FocusNextMenuBarItem

Focus the next top-level menu bar item (used when pressing arrow keys on menu bar items).

FocusPrevMenuBarItem

Focus the previous top-level menu bar item (used when pressing arrow keys on menu bar items).