Struct vizia::views::Label

pub struct Label { /* private fields */ }

A label used to display text.

Examples

Basic label

A label can be used to simply display some text on the screen.

Label::new(cx, "Text");

Label bound to data

A label can be bound to data using a lens which automatically updates the text whenever the underlying data changes.

#[derive(Lens)]
struct AppData {
    text: String,
}

impl Model for AppData {}

AppData {
    text: String::from("Text"),
}
.build(cx);

Label::new(cx, AppData::text);

Label with text wrapping

A label automatically wraps the text if it doesn’t fit inside of the width of the label.

Label::new(
    cx,
    "This is a really long text to showcase the text wrapping support of a label.",
)
.width(Pixels(100.0));

Label without text wrapping

A label can also be configured to never wrap the text by using the text_wrap method.

Label::new(
    cx,
    "This is a really long text to showcase disabled text wrapping of a label.",
)
.width(Pixels(100.0))
.text_wrap(false);

Label for a button

A label can also be used inside of a button to be able to add text to it.

Button::new(cx, |_| {}, |cx| Label::new(cx, "Text"));

Implementations

impl Label

pub fn new<T>(cx: &mut Context, text: impl Res<T> + Clone) -> Handle<'_, Label>
where T: ToStringLocalized,

Creates a new label.

Examples

Label::new(cx, "Text");

pub fn rich<T>( cx: &mut Context, text: impl Res<T> + Clone, children: impl Fn(&mut Context), ) -> Handle<'_, Label>
where T: ToStringLocalized,