vizia::window

Enum WindowEvent

pub enum WindowEvent {

    WindowClose,
    WindowMoved(WindowPosition),
    Drop(DropData),
    MouseDoubleClick(MouseButton),
    MouseTripleClick(MouseButton),
    MouseDown(MouseButton),
    MouseUp(MouseButton),
    Press {
        mouse: bool,
    },
    PressDown {
        mouse: bool,
    },
    MouseMove(f32, f32),
    MouseScroll(f32, f32),
    MouseOver,
    MouseOut,
    MouseEnter,
    MouseLeave,
    FocusIn,
    FocusOut,
    FocusVisibility(bool),
    WindowFocused(bool),
    CharInput(char),
    ImeActivate(bool),
    ImeCommit(String),
    ImePreedit(String, Option<(usize, usize)>),
    SetImeCursorArea((u32, u32), (u32, u32)),
    KeyDown(Code, Option<Key>),
    KeyUp(Code, Option<Key>),
    ThemeChanged(ThemeMode),
    SetCursor(CursorIcon),
    GrabCursor(bool),
    SetCursorPosition(u32, u32),
    SetTitle(String),
    SetSize(WindowSize),
    SetPosition(WindowPosition),
    SetMaxSize(Option<WindowSize>),
    SetMinSize(Option<WindowSize>),
    SetResizable(bool),
    SetMinimized(bool),
    SetMaximized(bool),
    SetVisible(bool),
    SetDecorations(bool),
    SetAlwaysOnTop(bool),
    MouseCaptureEvent,
    MouseCaptureOutEvent,
    GeometryChanged(GeoChanged),
    Redraw,
    Restyle,
    Relayout,
    FocusNext,
    FocusPrev,
    Debug(String),
    ActionRequest(ActionRequest),
    ReloadStyles,
    SetEnabled(bool),
    DragWindow,
    Destroyed,

}

Events generated by the application in response to OS events as well as events that can be used to set properties of the window.

Variants

WindowClose

Emitted when a window is closed. Can also be emitted by a view or model to close the window.

WindowMoved(WindowPosition)

Emitted when the window is moved.

Drop(DropData)

Emitted when a file is dragged and then dropped onto the window.

MouseDoubleClick(MouseButton)

Emitted when a mouse button is double clicked.

MouseTripleClick(MouseButton)

Emitted when a mouse button is triple clicked

MouseDown(MouseButton)

Emitted when a mouse button is pressed

MouseUp(MouseButton)

Emitted when a mouse button is released.

Press

Emitted when the primary mouse button or trigger key is pressed and then released on a view

Fields

mouse: bool

Whether the press event was triggered by the mouse.

PressDown

Emitted when the primary mouse button or trigger key is pressed on a view

Fields

mouse: bool

Whether the press down event was triggered by the mouse.

MouseMove(f32, f32)

Emitted when the mouse cursor is moved

MouseScroll(f32, f32)

Emitted when the mouse scroll wheel is scrolled.

MouseOver

Emitted when the mouse cursor enters the bounding box of an entity.

MouseOut

Emitted when the mouse cursor leaves the bounding box of an entity.

MouseEnter

Emitted when the mouse cursor enters an entity.

MouseLeave

Emitted when the mouse cursor leaves an entity.

FocusIn

Emitted when an entity gains keyboard focus.

FocusOut

Emitted when an entity loses keyboard focus.

FocusVisibility(bool)

Emitted when an entity’s focus visibility has changed.

WindowFocused(bool)

Emitted when the window gains or loses focus

CharInput(char)

Emitted when a character is typed.

ImeActivate(bool)

Emitted when the input method (IME) is activated or deactivated.

ImeCommit(String)

Emitted when an input method (IME) commits a string.

ImePreedit(String, Option<(usize, usize)>)

Emitted when an input method (IME) changes the preedit string.

SetImeCursorArea((u32, u32), (u32, u32))

Emitted when the input method (IME) area needs to be updated.

KeyDown(Code, Option<Key>)

Emitted when a keyboard key is pressed.

KeyUp(Code, Option<Key>)

Emitted when a keyboard key is released.

ThemeChanged(ThemeMode)

Emited when the system window theme has changed.

SetCursor(CursorIcon)

Sets the mouse cursor icon.

GrabCursor(bool)

Grabs the mouse cursor, preventing it from leaving the window.

SetCursorPosition(u32, u32)

Sets the (x,y) position of the mouse cursor in window coordinates.

SetTitle(String)

Sets the title of the window.

SetSize(WindowSize)

Sets the size of the window.

SetPosition(WindowPosition)

Sets the position of the window.

SetMaxSize(Option<WindowSize>)

Sets the maximum size of the window.

SetMinSize(Option<WindowSize>)

Sets the minimum size of the window.

SetResizable(bool)

Sets whether the window is resizable.

SetMinimized(bool)

Sets whether the window is minimized.

SetMaximized(bool)

Sets whether the window is maximized.

SetVisible(bool)

Sets whether the window is visible.

SetDecorations(bool)

Sets whether the window has decorations.

SetAlwaysOnTop(bool)

Sets whether the window remains on top of other windows.

MouseCaptureEvent

Emitted when mouse events have been captured.

MouseCaptureOutEvent

Emitted when mouse events have been released.

GeometryChanged(GeoChanged)

Emitted when an entity changes position or size.

Redraw

Requests a redraw of the window contents.

Restyle

Request a restyle.

Relayout

Requests a relayout.

FocusNext

Move keyboard focus to the next navigable view.

FocusPrev

Move keyboard focus to the previous navigable view.

Debug(String)

Prints the debug message to the console.

ActionRequest(ActionRequest)

Represents an action requested by an accessibility technology.

ReloadStyles

Reloads all application stylesheets.

SetEnabled(bool)

Enables or disabled mouse and keyboard input to the window.

DragWindow

Puts the window in a drag state.

Destroyed

Emitted when the window is destroyed.