Crate vizia

Vizia

Modules

binding

Data binding provides a way to link views to model data so that view properties update when data changes.

context

Context types for retained state, used during view building, event handling, and drawing.

environment

A model for system specific state which can be accessed by any model or view.

events

Events for communicating state change to views and models.

fonts

icons

A collection of codepoints for built-in icons.

image

Contains types and functions used for loading and manipulating images. This is a re-export of image.

input

Types used for handling input events such as mouse and keyboard.

layout

Vizia uses morphorm for layout.

localization

Provides types for adapting an application to a particular language or regional peculiarities.

model

Models are used to store application data and can be bound to by views to visually display the data.

modifiers

Methods on views for changing their properties or for adding actions.

resource

Resource management for fonts, themes, images, and translations.

style

Styling determines the appearance of a view.

util

vg

Contains types and functions used for custom drawing within views. This is a re-export of femtovg.

view

Views are used to visually present model data and to act as controls which, when interacted with, send events to mutate model data.

views

Built-in views provided by vizia.

window

Window management and system events.

Macros

include_style

Structs

Application

Creating a new application creates a root Window and a Context. Views declared within the closure passed to Application::new() are added to the context and rendered into the root window.

Constants

ICON_123

ICON_ADJUSTMENTS_ALT

ICON_ARROW_AUTOFIT_CONTENT

ICON_ARROW_AUTOFIT_DOWN

ICON_ARROW_AUTOFIT_HEIGHT

ICON_ARROW_AUTOFIT_LEFT

ICON_ARROW_AUTOFIT_RIGHT

ICON_ARROW_AUTOFIT_UP

ICON_ARROW_AUTOFIT_WIDTH

ICON_ARROW_BACK

ICON_ARROW_BACK_UP

ICON_ARROW_BACK_UP_DOUBLE

ICON_ARROW_BADGE_DOWN

ICON_ARROW_BADGE_LEFT

ICON_ARROW_BADGE_RIGHT

ICON_ARROW_BADGE_UP

ICON_ARROW_BAR_DOWN

ICON_ARROW_BAR_LEFT

ICON_ARROW_BAR_RIGHT

ICON_ARROW_BAR_TO_DOWN

ICON_ARROW_BAR_TO_LEFT

ICON_ARROW_BAR_TO_RIGHT

ICON_ARROW_BAR_TO_UP

ICON_ARROW_BAR_UP

ICON_ARROW_BEAR_LEFT

ICON_ARROW_BEAR_LEFT_2

ICON_ARROW_BEAR_RIGHT

ICON_ARROW_BEAR_RIGHT_2

ICON_ARROW_BIG_DOWN

ICON_ARROW_BIG_DOWN_LINE

ICON_ARROW_BIG_DOWN_LINES

ICON_ARROW_BIG_LEFT

ICON_ARROW_BIG_LEFT_LINE

ICON_ARROW_BIG_LEFT_LINES

ICON_ARROW_BIG_RIGHT

ICON_ARROW_BIG_RIGHT_LINE

ICON_ARROW_BIG_RIGHT_LINES

ICON_ARROW_BIG_UP

ICON_ARROW_BIG_UP_LINE

ICON_ARROW_BIG_UP_LINES

ICON_BAT

ICON_BOX_MULTIPLE_0

ICON_BOX_MULTIPLE_1

ICON_BOX_MULTIPLE_2

ICON_BOX_MULTIPLE_3

ICON_BOX_MULTIPLE_4

ICON_BOX_MULTIPLE_5

ICON_BOX_MULTIPLE_6

ICON_BOX_MULTIPLE_7

ICON_BOX_MULTIPLE_8

ICON_BOX_MULTIPLE_9

ICON_CAT

ICON_CHECK

ICON_CHEVRON_DOWN

ICON_CHEVRON_DOWN_LEFT

ICON_CHEVRON_DOWN_RIGHT

ICON_CHEVRON_LEFT

ICON_CHEVRON_RIGHT

ICON_CHEVRON_UP

ICON_CHEVRON_UP_LEFT

ICON_CHEVRON_UP_RIGHT

ICON_CIRCLE_NUMBER_0

ICON_CIRCLE_NUMBER_1

ICON_CIRCLE_NUMBER_2

ICON_CIRCLE_NUMBER_3

ICON_CIRCLE_NUMBER_4

ICON_CIRCLE_NUMBER_5

ICON_CIRCLE_NUMBER_6

ICON_CIRCLE_NUMBER_7

ICON_CIRCLE_NUMBER_8

ICON_CIRCLE_NUMBER_9

ICON_CLIPBOARD

ICON_COMMAND

ICON_COPY

ICON_CURSOR_TEXT

ICON_CUT

ICON_DEER

ICON_DOG

ICON_EDIT

ICON_EDIT_CIRCLE

ICON_EDIT_CIRCLE_OFF

ICON_EDIT_OFF

ICON_EYE

ICON_EYE_OFF

ICON_FILE_PENCIL

ICON_FISH

ICON_FISH_BONE

ICON_FISH_OFF

ICON_HEXAGON_NUMBER_0

ICON_HEXAGON_NUMBER_1

ICON_HEXAGON_NUMBER_2

ICON_HEXAGON_NUMBER_3

ICON_HEXAGON_NUMBER_4

ICON_HEXAGON_NUMBER_5

ICON_HEXAGON_NUMBER_6

ICON_HEXAGON_NUMBER_7

ICON_HEXAGON_NUMBER_8

ICON_HEXAGON_NUMBER_9

ICON_MINUS

ICON_MOON

ICON_PENCIL

ICON_PENCIL_MINUS

ICON_PENCIL_OFF

ICON_PENCIL_PLUS

ICON_PIG

ICON_PIG_MONEY

ICON_PIG_OFF

ICON_PLAYER_PAUSE

ICON_PLAYER_PAUSE_FILLED

ICON_PLAYER_PLAY

ICON_PLAYER_PLAY_FILLED

ICON_PLAYER_RECORD

ICON_PLAYER_RECORD_FILLED

ICON_PLAYER_SKIP_BACK

ICON_PLAYER_SKIP_BACK_FILLED

ICON_PLAYER_SKIP_FORWARD

ICON_PLAYER_SKIP_FORWARD_FILLED

ICON_PLAYER_STOP

ICON_PLAYER_STOP_FILLED

ICON_PLUS

ICON_POINTER

ICON_REPEAT

ICON_REPEAT_OFF

ICON_SEARCH

ICON_SETTINGS_AUTOMATION

ICON_SHARE

ICON_SHARE_2

ICON_SHARE_3

ICON_SLICE

ICON_SPIDER

ICON_STAR_FILLED

ICON_SUN

ICON_TRASH

ICON_X

ROBOTO_BOLD

ROBOTO_ITALIC

ROBOTO_REGULAR

TABLER_ICONS

Functions

log

One very small function for abstracting debugging between web and desktop programming. On the desktop, it will print to stdout, and on the web, it will print to the console log.