Trait vizia::binding::Res

pub trait Res<T> {
    // Required methods
    fn get_val(&self, cx: &Context) -> T;
    fn set_or_bind<F>(&self, cx: &mut Context, entity: Entity, closure: F)
       where F: 'static + Clone + Fn(&mut EventContext<'_>, T);

    // Provided method
    fn get_val_fallible(&self, cx: &Context) -> Option<T> { ... }
}

A trait which allows passing a value or a lens to a view or modifier.

For example, the Label view constructor takes a type which implements Res<T> where T implements ToString. This allows the user to pass a type which implements ToString, such as String or &str, or a lens to a type which implements ToString.

Required Methods

fn get_val(&self, cx: &Context) -> T

fn set_or_bind<F>(&self, cx: &mut Context, entity: Entity, closure: F)where F: 'static + Clone + Fn(&mut EventContext<'_>, T),

Provided Methods

fn get_val_fallible(&self, cx: &Context) -> Option<T>