Struct vizia::context::backend::BackendContext

pub struct BackendContext<'a>(pub &'a mut Context);

Tuple Fields
0: &'a mut Context
Implementations

impl<'a> BackendContext<'a>

pub fn new(cx: &'a mut Context) -> BackendContext<'a>

pub fn views(
&mut self
) -> &mut HashMap<Entity, Box<dyn ViewHandler + 'static, Global>, BuildHasherDefault<FnvHasher>>

pub fn style(&mut self) -> &mut Style

pub fn cache(&mut self) -> &mut CachedData

pub fn modifiers(&mut self) -> &mut Modifiers

pub fn add_main_window(
&mut self,
window_description: &WindowDescription,
canvas: Canvas<OpenGl>,
scale_factor: f32
)

pub fn environment(&self) -> &Environment

pub fn context(&mut self) -> &mut Context

pub fn remove_all_children(&mut self)

pub fn draw(&mut self)

pub fn load_images(&mut self)

pub fn set_current(&mut self, e: Entity)

Set the current entity. This is useful in user code when you’re performing black magic and want to trick other parts of the code into thinking you’re processing some other part of the tree.

pub fn with_current(&mut self, e: Entity, f: impl FnOnce(&mut Context))

Temporarily sets the current entity, calls the provided closure, and then resets the current entity back to previous.

pub fn set_event_proxy(&mut self, proxy: Box<dyn EventProxy + 'static, Global>)

You should not call this method unless you are writing a windowing backend, in which case you should consult the existing windowing backends for usage information.

pub fn set_clipboard_provider(
&mut self,
clipboard: Box<dyn ClipboardProvider + 'static, Global>
)

You should not call this method unless you are writing a windowing backend, in which case you should consult the existing windowing backends for usage information.

pub fn send_event(&mut self, event: Event)

Send an event with custom origin and propagation information.

pub fn has_queued_events(&self) -> bool

Check whether there are any events in the queue waiting for the next event dispatch cycle.

pub fn synchronize_fonts(&mut self)

Ensure all FontOrId entires are loaded into the contexts and become Ids.

pub fn text_context(&mut self) -> &mut TextContext

pub fn process_data_updates(&mut self)

For each binding or data observer, check if its data has changed, and if so, rerun its builder/body.

pub fn process_style_updates(&mut self)

pub fn process_visual_updates(&mut self)

Massages the style system until everything is coherent

pub fn dispatch_system_event(&mut self, event: WindowEvent)

This method is in charge of receiving raw WindowEvents and dispatching them to the appropriate points in the tree.

pub fn dispatch_direct_or_up(
&mut self,
event: WindowEvent,
direct: Entity,
up: Entity,
root: bool
)