Struct vizia::context::Context

pub struct Context {
    pub ignore_default_theme: bool,
    pub window_has_focus: bool,
    /* private fields */
}

The main storage and control object for a Vizia application.

Fields
ignore_default_theme: boolwindow_has_focus: bool
Implementations

impl Context

pub fn new(window_size: WindowSize, user_scale_factor: f64) -> Context

Creates a new context.

pub fn current(&self) -> Entity

The “current” entity, generally the entity which is currently being built or the entity which is currently having an event dispatched to it.

pub fn with_current<T>( &mut self, e: Entity, f: impl FnOnce(&mut Context) -> T ) -> T

Makes the above black magic more explicit

pub fn environment(&self) -> &Environment

Returns a reference to the Environment model.

pub fn window_size(&self) -> WindowSize

The window’s size in logical pixels, before user_scale_factor() gets applied to it. If this value changed during a frame then the window will be resized and a WindowEvent::GeometryChanged will be emitted.

pub fn user_scale_factor(&self) -> f64

A scale factor used for uniformly scaling the window independently of any HiDPI scaling. window_size gets multplied with this factor to get the actual logical window size. If this changes during a frame, then the window will be resized at the end of the frame and a WindowEvent::GeometryChanged will be emitted. This can be initialized using WindowDescription::user_scale_factor.

pub fn scale_factor(&self) -> f32

Returns the scale factor of the display.

pub fn needs_redraw(&mut self)

Mark the application as needing to rerun the draw method

pub fn needs_restyle(&mut self)

Mark the application as needing to recompute view styles

pub fn needs_relayout(&mut self)

Mark the application as needing to rerun layout computations

pub fn focus_with_visibility(&mut self, focus_visible: bool)

Sets application focus to the current entity with the specified focus visiblity

pub fn focus(&mut self)

Sets application focus to the current entity using the previous focus visibility

pub fn add_listener<F, W>(&mut self, listener: F)where W: View, F: 'static + Fn(&mut W, &mut EventContext<'_>, &mut Event),

Add a listener to an entity.

A listener can be used to handle events which would not normally propagate to the entity. For example, mouse events when a different entity has captured them. Useful for things like closing a popup when clicking outside of its bounding box.

pub fn add_global_listener<F>(&mut self, listener: F)where F: 'static + Fn(&mut EventContext<'_>, &mut Event),

Adds a global listener to the application.

Global listeners have the first opportunity to handle every event that is sent in an application. They will never be removed. If you need a listener tied to the lifetime of a view, use add_listener.

pub fn add_fonts_mem(&mut self, data: &[&[u8]])

Add a font from memory to the application.

The include_bytes!() macro can be used to embed a font into the application binary.

Example

cx.add_fonts_mem(&[include_bytes!("../../resources/fonts/Roboto-Regular.ttf")]);

pub fn set_language(&mut self, lang: LanguageIdentifier)

Sets the language used by the application for localization.

pub fn set_default_font(&mut self, names: &[&str])

Sets the global default font for the application.

pub fn add_stylesheet(&mut self, style: impl IntoCssStr) -> Result<(), Error>

pub fn remove_user_themes(&mut self)

Remove all user themes from the application.

pub fn add_animation(&mut self, animation: AnimationBuilder<'_>) -> Animation

pub fn set_image_loader<F>(&mut self, loader: F)where F: 'static + Fn(&mut ResourceContext<'_>, &str),

pub fn add_translation(&mut self, lang: LanguageIdentifier, ftl: impl ToString)

pub fn load_image( &mut self, path: &str, image: DynamicImage, policy: ImageRetentionPolicy )

pub fn spawn<F>(&self, target: F)where F: 'static + Send + FnOnce(&mut ContextProxy),

pub fn get_proxy(&self) -> ContextProxy

pub fn resolve_entity_identifier(&self, identity: &str) -> Option<Entity>

Finds the entity that identifier identifies