Struct vizia::handle::Handle

pub struct Handle<'a, V> {
    pub entity: Entity,
    pub p: PhantomData<V>,
    pub cx: &'a mut Context,
}

A handle to a view which has been already built into the tree.

This type is part of the prelude.

Fields
entity: Entityp: PhantomData<V>cx: &'a mut Context
Implementations

impl<'a, V> Handle<'a, V>

pub fn entity(&self) -> Entity

pub fn ignore(self) -> Handle<'a, V>

pub fn lock_focus_to_within(self) -> Handle<'a, V>

Stop the user from tabbing out of a subtree, which is useful for modal dialogs.

pub fn modify<F>(self, f: F) -> Handle<'a, V>where
F: FnOnce(&mut V),
V: 'static,

pub fn on_build<F>(self, callback: F) -> Handle<'a, V>where
F: Fn(&mut EventContext<'_>),

Callback which is run when the view is built/rebuilt.

pub fn bind<L, F>(self, lens: L, closure: F) -> Handle<'a, V>where
L: Lens,
<L as Lens>::Target: Data,
F: 'static + Fn(Handle<'_, V>, L),

impl Handle<'_, Checkbox>

pub fn on_toggle<F>(self, callback: F) -> Handle<'_, Checkbox>where
F: 'static + Fn(&mut EventContext<'_>),

Set the callback triggered when the checkbox is pressed.

Examples

Checkbox::new(cx, AppData::value).on_toggle(|cx| cx.emit(AppEvent::ToggleValue));

impl<'a, L> Handle<'a, Knob<L>>where
L: Lens<Target = f32>,

pub fn on_changing<F>(self, callback: F) -> Handle<'a, Knob<L>>where
F: 'static + Fn(&mut EventContext<'_>, f32),

impl Handle<'_, TickKnob>

pub fn value<L>(self, lens: L) -> Handle<'_, TickKnob>where
L: Lens<Target = f32>,

impl Handle<'_, ArcTrack>

pub fn value<L>(self, lens: L) -> Handle<'_, ArcTrack>where
L: Lens<Target = f32>,

impl Handle<'_, Label>

pub fn describing(
self,
entity_identifier: impl Into<String>
) -> Handle<'_, Label>

Which form element does this label describe.

Examples

Checkbox::new(cx, AppData::value).on_toggle(|cx| cx.emit(AppEvent::ToggleValue)).id("checkbox_identifier");
Label::new(cx, "hello").describing("checkbox_identifier");

impl<L, T> Handle<'_, List<L, T>>where
L: Lens<Target = Vec<T, Global>>,
T: Data,

pub fn on_increment<F>(self, callback: F) -> Handle<'_, List<L, T>>where
F: 'static + Fn(&mut EventContext<'_>),

pub fn on_decrement<F>(self, callback: F) -> Handle<'_, List<L, T>>where
F: 'static + Fn(&mut EventContext<'_>),

pub fn on_clear<F>(self, callback: F) -> Handle<'_, List<L, T>>where
F: 'static + Fn(&mut EventContext<'_>),

impl<'a, L> Handle<'a, Popup<L>>where
L: Lens,
<L as Lens>::Target: Clone + Into<bool>,

pub fn on_blur<F>(self, f: F) -> Handle<'a, Popup<L>>where
F: 'static + Fn(&mut EventContext<'_>),

Registers a callback for when the user clicks off of the popup, usually with the intent of closing it.

impl Handle<'_, RadioButton>

pub fn on_select<F>(self, callback: F) -> Handle<'_, RadioButton>where
F: 'static + Fn(&mut EventContext<'_>),

Set the callback triggered when the radio button is selected.

Examples

RadioButton::new(cx, AppData::value).on_select(|cx| cx.emit(AppEvent::ToggleValue));

impl<L> Handle<'_, Slider<L>>where
L: Lens,

pub fn on_changing<F>(self, callback: F) -> Handle<'_, Slider<L>>where
F: 'static + Fn(&mut EventContext<'_>, f32),

Sets the callback triggered when the slider value is changing (dragging).

Takes a closure which triggers when the slider value is changing, either by pressing the track or dragging the thumb along the track.

Example

Slider::new(cx, AppData::value)
    .on_changing(|cx, value| {
        println!("Slider on_changing: {}", value);
    });

pub fn range(self, range: Range<f32>) -> Handle<'_, Slider<L>>

Sets the range of the slider.

If the bound data is outside of the range then the slider will clip to min/max of the range.

Example

Slider::new(cx, AppData::value)
    .range(-20.0..50.0)
    .on_changing(|cx, value| {
        println!("Slider on_changing: {}", value);
    });

pub fn keyboard_fraction(self, keyboard_fraction: f32) -> Handle<'_, Slider<L>>

Sets the fraction of a slider that a press of an arrow key will change.

Example

Slider::new(cx, AppData::value)
    .keyboard_fraction(0.05)
    .on_changing(|cx, value| {
        println!("Slider on_changing: {}", value);
    });

impl<'a, L> Handle<'a, Textbox<L>>where
L: Lens,

pub fn on_edit<F>(self, callback: F) -> Handle<'a, Textbox<L>>where
F: 'static + Fn(&mut EventContext<'_>, String) + Send + Sync,

pub fn on_submit<F>(self, callback: F) -> Handle<'a, Textbox<L>>where
F: 'static + Fn(&mut EventContext<'_>, String, bool) + Send + Sync,