Struct vizia::handle::Handle

pub struct Handle<'a, T> {
    pub entity: Entity,
    pub p: PhantomData<T>,
    pub cx: &'a mut Context,
}

A handle to a view which has been already built into the tree.

This type is part of the prelude.

Fields
entity: Entityp: PhantomData<T>cx: &'a mut Context
Implementations

impl<'a, T> Handle<'a, T>

pub fn entity(&self) -> Entity

pub fn ignore(self) -> Handle<'a, T>

pub fn modify<F>(self, f: F) -> Handle<'a, T> where
F: FnOnce(&mut T),
T: 'static,

pub fn on_build<F>(self, callback: F) -> Handle<'a, T> where
F: Fn(&mut Context),

Callback which is run when the view is built/rebuilt

pub fn bind<L, F>(self, lens: L, closure: F) -> Handle<'a, T> where
L: Lens,
F: 'static + Fn(Handle<'_, T>, L),
<L as Lens>::Target: Data,

pub fn id(self, id: &str) -> Handle<'a, T>

pub fn cursor(self, cursor_icon: CursorIcon) -> Handle<'a, T>

pub fn class(self, name: &str) -> Handle<'a, T>

pub fn toggle_class(self, name: &str, applied: impl Res<bool>) -> Handle<'a, T>

pub fn font(self, font_name: &str) -> Handle<'a, T>

pub fn checked(self, state: impl Res<bool>) -> Handle<'a, T>

pub fn disabled(self, state: impl Res<bool>) -> Handle<'a, T>

pub fn text<U>(self, value: impl Res<U>) -> Handle<'a, T> where
U: ToString,

pub fn image<U>(self, value: impl Res<U>) -> Handle<'a, T> where
U: ToString,

pub fn z_order(self, value: i32) -> Handle<'a, T>

pub fn overflow(self, value: Overflow) -> Handle<'a, T>

pub fn display<U>(self, value: impl Res<U>) -> Handle<'a, T> where
U: Clone + Into<Display>,

pub fn visibility<U>(self, value: impl Res<U>) -> Handle<'a, T> where
U: Clone + Into<Visibility>,

pub fn hoverable(self, state: bool) -> Handle<'a, T>

pub fn focusable(self, state: bool) -> Handle<'a, T>

pub fn keyboard_navigatable(self, state: bool) -> Handle<'a, T>

pub fn child_space(self, value: Units) -> Handle<'a, T>

pub fn border_radius(self, value: Units) -> Handle<'a, T>

pub fn space(self, value: Units) -> Handle<'a, T>

pub fn size(self, value: Units) -> Handle<'a, T>

pub fn min_size(self, value: Units) -> Handle<'a, T>

pub fn max_size(self, value: Units) -> Handle<'a, T>

pub fn color(self, color: Color) -> Handle<'a, T>

pub fn grid_rows(self, rows: Vec<Units, Global>) -> Handle<'a, T>

pub fn grid_cols(self, cols: Vec<Units, Global>) -> Handle<'a, T>

pub fn background_color(self, value: impl Res<Color>) -> Handle<'a, T>

pub fn background_image(self, value: impl Res<String>) -> Handle<'a, T>

pub fn layout_type(self, value: impl Res<LayoutType>) -> Handle<'a, T>

pub fn position_type(self, value: impl Res<PositionType>) -> Handle<'a, T>

pub fn left(self, value: impl Res<Units>) -> Handle<'a, T>

pub fn right(self, value: impl Res<Units>) -> Handle<'a, T>

pub fn top(self, value: impl Res<Units>) -> Handle<'a, T>

pub fn bottom(self, value: impl Res<Units>) -> Handle<'a, T>

pub fn width(self, value: impl Res<Units>) -> Handle<'a, T>

pub fn height(self, value: impl Res<Units>) -> Handle<'a, T>

pub fn min_width(self, value: impl Res<Units>) -> Handle<'a, T>

pub fn max_width(self, value: impl Res<Units>) -> Handle<'a, T>

pub fn min_height(self, value: impl Res<Units>) -> Handle<'a, T>

pub fn max_height(self, value: impl Res<Units>) -> Handle<'a, T>

pub fn min_left(self, value: impl Res<Units>) -> Handle<'a, T>

pub fn max_left(self, value: impl Res<Units>) -> Handle<'a, T>

pub fn min_right(self, value: impl Res<Units>) -> Handle<'a, T>

pub fn max_right(self, value: impl Res<Units>) -> Handle<'a, T>

pub fn min_top(self, value: impl Res<Units>) -> Handle<'a, T>

pub fn max_top(self, value: impl Res<Units>) -> Handle<'a, T>

pub fn min_bottom(self, value: impl Res<Units>) -> Handle<'a, T>

pub fn max_bottom(self, value: impl Res<Units>) -> Handle<'a, T>

pub fn child_left(self, value: impl Res<Units>) -> Handle<'a, T>

pub fn child_right(self, value: impl Res<Units>) -> Handle<'a, T>

pub fn child_top(self, value: impl Res<Units>) -> Handle<'a, T>

pub fn child_bottom(self, value: impl Res<Units>) -> Handle<'a, T>

pub fn row_between(self, value: impl Res<Units>) -> Handle<'a, T>

pub fn col_between(self, value: impl Res<Units>) -> Handle<'a, T>

pub fn row_index(self, value: impl Res<usize>) -> Handle<'a, T>

pub fn row_span(self, value: impl Res<usize>) -> Handle<'a, T>

pub fn col_index(self, value: impl Res<usize>) -> Handle<'a, T>

pub fn col_span(self, value: impl Res<usize>) -> Handle<'a, T>

pub fn border_width(self, value: impl Res<Units>) -> Handle<'a, T>

pub fn border_color(self, value: impl Res<Color>) -> Handle<'a, T>

pub fn font_size(self, value: impl Res<f32>) -> Handle<'a, T>

pub fn text_selection(self, value: impl Res<Selection>) -> Handle<'a, T>

pub fn caret_color(self, value: impl Res<Color>) -> Handle<'a, T>

pub fn selection_color(self, value: impl Res<Color>) -> Handle<'a, T>

pub fn text_wrap(self, value: impl Res<bool>) -> Handle<'a, T>

pub fn rotate(self, value: impl Res<f32>) -> Handle<'a, T>

pub fn translate(self, value: impl Res<(f32, f32)>) -> Handle<'a, T>

pub fn scale(self, value: impl Res<(f32, f32)>) -> Handle<'a, T>

pub fn border_shape_top_left(
self,
value: impl Res<BorderCornerShape>
) -> Handle<'a, T>

pub fn border_shape_top_right(
self,
value: impl Res<BorderCornerShape>
) -> Handle<'a, T>

pub fn border_shape_bottom_left(
self,
value: impl Res<BorderCornerShape>
) -> Handle<'a, T>

pub fn border_shape_bottom_right(
self,
value: impl Res<BorderCornerShape>
) -> Handle<'a, T>

pub fn border_radius_top_left(self, value: impl Res<Units>) -> Handle<'a, T>

pub fn border_radius_top_right(self, value: impl Res<Units>) -> Handle<'a, T>

pub fn border_radius_bottom_left(self, value: impl Res<Units>) -> Handle<'a, T>

pub fn border_radius_bottom_right(self, value: impl Res<Units>) -> Handle<'a, T>

impl<'_> Handle<'_, Checkbox>

pub fn on_toggle<F>(self, callback: F) -> Handle<'_, Checkbox> where
F: 'static + Fn(&mut Context),

Set the callback triggered when the checkbox is pressed.

Examples

Checkbox::new(cx, AppData::value).on_toggle(|cx| cx.emit(AppEvent::ToggleValue));

impl<'a, L> Handle<'a, Knob<L>> where
L: Lens<Target = f32>,

pub fn on_changing<F>(self, callback: F) -> Handle<'a, Knob<L>> where
F: 'static + Fn(&mut Context, f32),

impl<'_> Handle<'_, TickKnob>

pub fn value<L>(self, lens: L) -> Handle<'_, TickKnob> where
L: Lens<Target = f32>,

impl<'_> Handle<'_, ArcTrack>

pub fn value<L>(self, lens: L) -> Handle<'_, ArcTrack> where
L: Lens<Target = f32>,

impl<'_, L, T> Handle<'_, List<L, T>> where
L: Lens<Target = Vec<T, Global>>,
T: Data,

pub fn on_increment<F>(self, callback: F) -> Handle<'_, List<L, T>> where
F: 'static + Fn(&mut Context),

pub fn on_decrement<F>(self, callback: F) -> Handle<'_, List<L, T>> where
F: 'static + Fn(&mut Context),

pub fn on_clear<F>(self, callback: F) -> Handle<'_, List<L, T>> where
F: 'static + Fn(&mut Context),

impl<'a, L> Handle<'a, Popup<L>> where
L: Lens,
<L as Lens>::Target: Clone,
<L as Lens>::Target: Into<bool>,

pub fn on_blur<F>(self, f: F) -> Handle<'a, Popup<L>> where
F: 'static + Fn(&mut Context),

Registers a callback for when the user clicks off of the popup, usually with the intent of closing it.

impl<'_> Handle<'_, RadioButton>

pub fn on_select<F>(self, callback: F) -> Handle<'_, RadioButton> where
F: 'static + Fn(&mut Context),

Set the callback triggered when the radio button is selected.

Examples

RadioButton::new(cx, AppData::value).on_select(|cx| cx.emit(AppEvent::ToggleValue));

impl<'_, L> Handle<'_, Slider<L>> where
L: Lens,

pub fn on_changing<F>(self, callback: F) -> Handle<'_, Slider<L>> where
F: 'static + Fn(&mut Context, f32),

Sets the callback triggered when the slider value is changing (dragging).

Takes a closure which triggers when the slider value is changing, either by pressing the track or dragging the thumb along the track.

Example

Slider::new(cx, AppData::value)
    .on_changing(|cx, value| {
        println!("Slider on_changing: {}", value);
    });

pub fn range(self, range: Range<f32>) -> Handle<'_, Slider<L>>

Sets the range of the slider.

If the bound data is outside of the range then the slider will clip to min/max of the range.

Example

Slider::new(cx, AppData::value)
    .range(-20.0..50.0)
    .on_changing(|cx, value| {
        println!("Slider on_changing: {}", value);
    });

impl<'a, L> Handle<'a, Textbox<L>> where
L: Lens,

pub fn on_edit<F>(self, callback: F) -> Handle<'a, Textbox<L>> where
F: 'static + Fn(&mut Context, String) + Send + Sync,

pub fn on_submit<F>(self, callback: F) -> Handle<'a, Textbox<L>> where
F: 'static + Fn(&mut Context, String, bool) + Send + Sync,