Struct vizia::input::Keymap

pub struct Keymap<T>where
    T: 'static + Clone + PartialEq<T> + Send + Sync,{ /* private fields */ }

A keymap that associates key chords with actions.

This is useful if you have an application that lets the user configure their key chords. It allows you to check if a particular action is pressed rather than the actual keys. The relationship between a key chord and an action is a many-to-many relationship.

Examples

First we need to create something that represents an action in our application. This is usually an enum.

#[derive(PartialEq, Copy, Clone)]
enum Action {
    One,
    Two,
    Three,
}

Now we can create a new keymap inside of our application and configure our key chords. We will bind Action::One to the key chord A, Action::Two to the key chord CTRL+B and Action::Three to the key chord CTRL+SHIFT+C. Every action has an associated callback function that gets triggered when the key chord is pressed.

let keymap = Keymap::from(vec![
    (KeyChord::new(Modifiers::empty(), Code::KeyA), KeymapEntry::new(Action::One, |_| println!("Action One"))),
    (KeyChord::new(Modifiers::CTRL, Code::KeyB), KeymapEntry::new(Action::Two, |_| println!("Action Two"))),
    (KeyChord::new(Modifiers::CTRL | Modifiers::SHIFT, Code::KeyC), KeymapEntry::new(Action::Three, |_| println!("Action Three"))),
]);

Implementations

impl<T> Keymap<T>where T: 'static + Clone + PartialEq<T> + Send + Sync,

pub fn new() -> Keymap<T>

Creates a new keymap.

Examples

let keymap = Keymap::<Action>::new();

pub fn pressed_actions( &self, cx: &Context, code: Code ) -> impl Iterator<Item = &KeymapEntry<T>>

Returns an iterator over every pressed keymap entry.

Examples

for entry in keymap.pressed_actions(cx, Code::KeyA) {
    println!("The action {:?} is being pressed!", entry.action());
};

pub fn export(&self) -> Vec<(&KeyChord, &KeymapEntry<T>), Global>

Exports all keymap entries and their associated key chords.

This is useful if you want to have a settings window and need to access every key chord and keymap entry of a keymap.

Examples

let actions_chords = keymap.export();

for (chord, entry) in actions_chords {
    println!("The key chord {:?} triggers the action {:?}!", chord, entry.action());
}