Trait vizia::modifiers::AbilityModifiers

pub trait AbilityModifiers: Modifiable {
    fn hoverable<U>(self, state: impl Res<U>) -> Self
    where
        U: Into<bool>,
    { ... }
    fn focusable<U>(self, state: impl Res<U>) -> Self
    where
        U: Into<bool>,
    { ... }
    fn navigable<U>(self, state: impl Res<U>) -> Self
    where
        U: Into<bool>,
    { ... }
}

Modifiers for changing the abilities of a view.

Provided Methods

fn hoverable<U>(self, state: impl Res<U>) -> Selfwhere
U: Into<bool>,

Sets whether the view can be hovered by the mouse.

Views which cannot be hovered will not receive mouse input events unless the view has captured the mouse input, see cx.capture().

fn focusable<U>(self, state: impl Res<U>) -> Selfwhere
U: Into<bool>,

Sets whether the view can be focused to receive keyboard input events.

fn navigable<U>(self, state: impl Res<U>) -> Selfwhere
U: Into<bool>,

Sets whether the view can be navigated to, i.e. focused, by the keyboard.

Navigating to a view with the keyboard gives the view keyboard focus and is typically done with tab and shift + tab key combinations.