Trait vizia::modifiers::TextModifiers

pub trait TextModifiers: Modifiable {

    // Provided methods
    fn text<T>(self, value: impl Res<T>) -> Self
       where T: ToStringLocalized { ... }
    fn font_family<U>(self, value: impl Res<U>) -> Self
       where U: Into<Vec<FamilyOwned>> { ... }
    fn font_weight<U>(self, value: impl Res<U>) -> Self
       where U: Into<FontWeight> { ... }
    fn font_slant<U>(self, value: impl Res<U>) -> Self
       where U: Into<FontSlant> { ... }
    fn font_width<U>(self, value: impl Res<U>) -> Self
       where U: Into<FontWidth> { ... }
    fn font_variation_settings<U>(self, value: impl Res<U>) -> Self
       where U: Into<Vec<FontVariation>> { ... }
    fn color<U>(self, value: impl Res<U>) -> Self
       where U: Clone + Into<Color> { ... }
    fn font_size<U>(self, value: impl Res<U>) -> Self
       where U: Into<FontSize> { ... }
    fn caret_color<U>(self, value: impl Res<U>) -> Self
       where U: Into<Color> { ... }
    fn selection_color<U>(self, value: impl Res<U>) -> Self
       where U: Into<Color> { ... }
    fn text_wrap<U>(self, value: impl Res<U>) -> Self
       where U: Into<bool> { ... }
    fn text_align<U>(self, value: impl Res<U>) -> Self
       where U: Into<TextAlign> { ... }
    fn text_overflow<U>(self, value: impl Res<U>) -> Self
       where U: Into<TextOverflow> { ... }
    fn line_clamp<U>(self, value: impl Res<U>) -> Self
       where U: Into<LineClamp> { ... }
    fn text_decoration_line<U>(self, value: impl Res<U>) -> Self
       where U: Into<TextDecorationLine> { ... }

}

Modifiers for changing the text properties of a view.

Provided Methods

fn text<T>(self, value: impl Res<T>) -> Self
where T: ToStringLocalized,

Sets the text content of the view.

fn font_family<U>(self, value: impl Res<U>) -> Self
where U: Into<Vec<FamilyOwned>>,

Sets the font that should be used by the view.

The font name refers to the name assigned when the font is added to context.

fn font_weight<U>(self, value: impl Res<U>) -> Self
where U: Into<FontWeight>,

Sets the font weight that should be used by the view.

fn font_slant<U>(self, value: impl Res<U>) -> Self
where U: Into<FontSlant>,

Sets the font style that should be used by the view.

fn font_width<U>(self, value: impl Res<U>) -> Self
where U: Into<FontWidth>,

Sets the font stretch that should be used by the view if the font supports it.

fn font_variation_settings<U>(self, value: impl Res<U>) -> Self
where U: Into<Vec<FontVariation>>,

Sets the font variation settings that should be used by the view.

fn color<U>(self, value: impl Res<U>) -> Self
where U: Clone + Into<Color>,

Sets the text color of the view.

fn font_size<U>(self, value: impl Res<U>) -> Self
where U: Into<FontSize>,

Sets the font size of the view.

fn caret_color<U>(self, value: impl Res<U>) -> Self
where U: Into<Color>,

Sets the ext caret color of the view.

fn selection_color<U>(self, value: impl Res<U>) -> Self
where U: Into<Color>,

Sets the color used to highlight selected text within the view.

fn text_wrap<U>(self, value: impl Res<U>) -> Self
where U: Into<bool>,

Sets whether the text of the view should be allowed to wrap.

fn text_align<U>(self, value: impl Res<U>) -> Self
where U: Into<TextAlign>,

Sets the horizontal alignment of text within the view.

fn text_overflow<U>(self, value: impl Res<U>) -> Self
where U: Into<TextOverflow>,

Sets the text overflow.

fn line_clamp<U>(self, value: impl Res<U>) -> Self
where U: Into<LineClamp>,

Sets the max number of .

fn text_decoration_line<U>(self, value: impl Res<U>) -> Self
where U: Into<TextDecorationLine>,

Sets the max number of .

Object Safety

This trait is not object safe.