Trait vizia::modifiers::TextModifiers

pub trait TextModifiers: Modifiable {
    fn text<U>(self, value: impl Res<U>) -> Self
    where
        U: ToString,
    { ... }
    fn font<U>(self, value: impl Res<U>) -> Self
    where
        U: Into<String>,
    { ... }
    fn color<U>(self, value: impl Res<U>) -> Self
    where
        U: Into<Color>,
    { ... }
    fn font_size(self, value: impl Res<f32>) -> Self { ... }
    fn text_selection<U>(self, value: impl Res<U>) -> Self
    where
        U: Into<Selection>,
    { ... }
    fn caret_color<U>(self, value: impl Res<U>) -> Self
    where
        U: Into<Color>,
    { ... }
    fn selection_color<U>(self, value: impl Res<U>) -> Self
    where
        U: Into<Color>,
    { ... }
    fn text_wrap<U>(self, value: impl Res<U>) -> Self
    where
        U: Into<bool>,
    { ... }
}

Modifiers for changing the text properties of a view.

Provided Methods

fn text<U>(self, value: impl Res<U>) -> Selfwhere
U: ToString,

Sets the text content of the view.

fn font<U>(self, value: impl Res<U>) -> Selfwhere
U: Into<String>,

Sets the font that should be used by the view.

The font name refers to the name assigned when the font is added to context.

fn color<U>(self, value: impl Res<U>) -> Selfwhere
U: Into<Color>,

Sets the text color of the view.

fn font_size(self, value: impl Res<f32>) -> Self

Sets the font size of the view.

fn text_selection<U>(self, value: impl Res<U>) -> Selfwhere
U: Into<Selection>,

Sets the text selection of the view.

fn caret_color<U>(self, value: impl Res<U>) -> Selfwhere
U: Into<Color>,

Sets the ext caret color of the view.

fn selection_color<U>(self, value: impl Res<U>) -> Selfwhere
U: Into<Color>,

Sets the color used to highlight selected text within the view.

fn text_wrap<U>(self, value: impl Res<U>) -> Selfwhere
U: Into<bool>,

Sets whether the text of the view should be allowed to wrap.