Trait vizia::modifiers::TextModifiers

pub trait TextModifiers: Modifiable {
    // Provided methods
    fn text<T>(self, value: impl Res<T>) -> Self
       where T: ToStringLocalized { ... }
    fn font_family<U>(self, value: impl Res<U>) -> Self
       where U: Into<Vec<FamilyOwned, Global>> { ... }
    fn font_weight<U>(self, value: impl Res<U>) -> Self
       where U: Into<FontWeight> { ... }
    fn font_style<U>(self, value: impl Res<U>) -> Self
       where U: Into<FontStyle> { ... }
    fn font_stretch<U>(self, value: impl Res<U>) -> Self
       where U: Into<FontStretch> { ... }
    fn color<U>(self, value: impl Res<U>) -> Self
       where U: Into<Color> { ... }
    fn font_size<U>(self, value: impl Res<U>) -> Self
       where U: Into<FontSize> { ... }
    fn caret_color<U>(self, value: impl Res<U>) -> Self
       where U: Into<Color> { ... }
    fn selection_color<U>(self, value: impl Res<U>) -> Self
       where U: Into<Color> { ... }
    fn text_wrap<U>(self, value: impl Res<U>) -> Self
       where U: Into<bool> { ... }
    fn text_align<U>(self, value: impl Res<U>) -> Self
       where U: Into<TextAlign> { ... }
}

Modifiers for changing the text properties of a view.

Provided Methods

fn text<T>(self, value: impl Res<T>) -> Selfwhere T: ToStringLocalized,

Sets the text content of the view.

fn font_family<U>(self, value: impl Res<U>) -> Selfwhere U: Into<Vec<FamilyOwned, Global>>,

Sets the font that should be used by the view.

The font name refers to the name assigned when the font is added to context.

fn font_weight<U>(self, value: impl Res<U>) -> Selfwhere U: Into<FontWeight>,

Sets the font weight that should be used by the view.

fn font_style<U>(self, value: impl Res<U>) -> Selfwhere U: Into<FontStyle>,

Sets the font style that should be used by the view.

fn font_stretch<U>(self, value: impl Res<U>) -> Selfwhere U: Into<FontStretch>,

Sets the font stretch that should be used by the view if the font supports it.

fn color<U>(self, value: impl Res<U>) -> Selfwhere U: Into<Color>,

Sets the text color of the view.

fn font_size<U>(self, value: impl Res<U>) -> Selfwhere U: Into<FontSize>,

Sets the font size of the view.

fn caret_color<U>(self, value: impl Res<U>) -> Selfwhere U: Into<Color>,

Sets the ext caret color of the view.

fn selection_color<U>(self, value: impl Res<U>) -> Selfwhere U: Into<Color>,

Sets the color used to highlight selected text within the view.

fn text_wrap<U>(self, value: impl Res<U>) -> Selfwhere U: Into<bool>,

Sets whether the text of the view should be allowed to wrap.

fn text_align<U>(self, value: impl Res<U>) -> Selfwhere U: Into<TextAlign>,

Sets the horizontal alignment of text within the view.