Struct vizia::prelude::Context

pub struct Context {
    pub views: HashMap<Entity, Box<dyn ViewHandler + 'static, Global>, BuildHasherDefault<FnvHasher>>,
    /* private fields */
}

The main storage and control object for a Vizia application.

This type is part of the prelude.

Fields
views: HashMap<Entity, Box<dyn ViewHandler + 'static, Global>, BuildHasherDefault<FnvHasher>>

TODO make this private when there’s no longer a need to mutate views after building

Implementations

impl Context

pub fn new() -> Context

pub fn tree(&mut self) -> &mut Tree

Access to the tree of entities

pub fn tree_ref(&self) -> &Tree

pub fn current(&self) -> Entity

The “current” entity, generally the entity which is currently being built or the entity which is currently having an event dispatched to it.

pub fn set_current(&mut self, e: Entity)

Set the current entity. This is useful in user code when you’re performing black magic and want to trick other parts of the code into thinking you’re processing some other part of the tree.

pub fn style(&mut self) -> &mut Style

The style storage for the application. Used to get properties set through inline style attributes or CSS.

pub fn style_ref(&self) -> &Style

pub fn cache(&mut self) -> &mut CachedData

The cache storage for the application. Used to get intermediate data computed during the layout and rendering processes.

pub fn cache_ref(&self) -> &CachedData

pub fn environment(&mut self) -> &mut Environment

pub fn text_context(&mut self) -> &mut TextContext

The current femtovg text context. Useful when measuring or rendering fonts.

pub fn mouse(&self) -> &MouseState

The current mouse state, i.e. the button and motion information.

pub fn resource_manager(&mut self) -> &mut ResourceManager

pub fn resource_manager_ref(&self) -> &ResourceManager

pub fn modifiers(&self) -> Modifiers

The current keyboard modifiers state.

pub fn modifiers_mut(&mut self) -> &mut Modifiers

pub fn need_redraw(&mut self)

Mark the application as needing to rerun the draw method

pub fn need_restyle(&mut self)

Mark the application as needing to recompute view styles

pub fn need_relayout(&mut self)

Mark the application as needing to rerun layout computations

pub fn capture(&mut self)

Causes mouse events to propagate to the current entity until released

pub fn release(&mut self)

Releases the mouse events capture

pub fn lock_cursor_icon(&mut self)

Prevents the cursor icon from changing until the lock is released

pub fn unlock_cursor_icon(&mut self)

Releases any cursor icon lock, allowing the cursor icon to be changed

pub fn is_cursor_icon_locked(&self) -> bool

Returns true if the cursor icon is locked

pub fn focus(&mut self)

Sets application focus to the current entity

pub fn set_active(&mut self, flag: bool)

Sets the active flag of the current entity

pub fn set_checked(&mut self, flag: bool)

Sets the checked flag of the current entity

pub fn set_selected(&mut self, flag: bool)

Sets the checked flag of the current entity

pub fn toggle_class(&mut self, class_name: &str, applied: bool)

pub fn is_disabled(&self) -> bool

Returns true if the current entity is disabled

pub fn is_over(&self) -> bool

Returns true if the mouse cursor is over the current entity

pub fn hovered(&self) -> Entity

pub fn captured(&self) -> Entity

pub fn set_event_proxy(&mut self, proxy: Box<dyn EventProxy + 'static, Global>)

You should not call this method unless you are writing a windowing backend, in which case you should consult the existing windowing backends for usage information.

pub fn set_clipboard_provider(
&mut self,
clipboard: Box<dyn ClipboardProvider + 'static, Global>
)

You should not call this method unless you are writing a windowing backend, in which case you should consult the existing windowing backends for usage information.

pub fn get_clipboard(
&mut self
) -> Result<String, Box<dyn Error + Send + Sync + 'static, Global>>

Get the contents of the system clipboard. This may fail for a variety of backend-specific reasons.

pub fn set_clipboard(
&mut self,
text: String
) -> Result<(), Box<dyn Error + Send + Sync + 'static, Global>>

Set the contents of the system clipboard. This may fail for a variety of backend-specific reasons.

pub fn has_pseudo_class(&self, entity: Entity, cls: PseudoClass) -> bool

Check whether an entity has a given pseudoclass, e.g. detecting if it’s checked, selected, etc.

pub fn remove_children(&mut self, entity: Entity)

pub fn remove(&mut self, entity: Entity)

pub fn emit<M>(&mut self, message: M) where
M: Message,

Send an event containing a message up the tree from the current entity.

pub fn emit_to<M>(&mut self, target: Entity, message: M) where
M: Message,

Send an event containing a message directly to a specified entity.

pub fn emit_custom(&mut self, event: Event)

Send an event with custom origin and propagation information.

pub fn has_queued_events(&self) -> bool

Check whether there are any events in the queue waiting for the next event dispatch cycle.

pub fn add_listener<F, W>(&mut self, listener: F) where
W: View,
F: 'static + Fn(&mut W, &mut Context, &mut Event),

Add a listener to an entity.

A listener can be used to handle events which would not normally propagate to the entity. For example, mouse events when a different entity has captured them. Useful for things like closing a popup when clicking outside of its bounding box.

pub fn add_font_mem(&mut self, name: &str, data: &[u8])

Add a font from memory to the application.

pub fn set_default_font(&mut self, name: &str)

Sets the global default font for the application.

pub fn synchronize_fonts(&mut self, canvas: &mut Canvas<OpenGl>)

Ensure all FontOrId entires are loaded into the contexts and become Ids.

pub fn add_theme(&mut self, theme: &str)

pub fn remove_user_themes(&mut self)

pub fn add_stylesheet(&mut self, path: &str) -> Result<(), Error>

pub fn has_animations(&self) -> bool

pub fn apply_animations(&mut self)

pub fn add_animation(&mut self, duration: Duration) -> AnimationBuilder<'_>

Adds a new property animation returning an animation builder

Example

Create an animation which animates the left property from 0 to 100 pixels in 5 seconds and play the animation on an entity:

let animation_id = cx.add_animation(instant::Duration::from_secs(5))
    .add_keyframe(0.0, |keyframe| keyframe.set_left(Pixels(0.0)))
    .add_keyframe(1.0, |keyframe| keyframe.set_left(Pixels(100.0)))
    .build();

pub fn play_animation(&mut self, animation: Animation)

pub fn reload_styles(&mut self) -> Result<(), Error>

pub fn set_image_loader<F>(&mut self, loader: F) where
F: 'static + Fn(&mut Context, &str),

pub fn get_image(&mut self, path: &str) -> &mut ImageOrId

pub fn add_image_observer(&mut self, path: &str, observer: Entity)

pub fn load_image(
&mut self,
path: String,
image: DynamicImage,
policy: ImageRetentionPolicy
)

pub fn evict_image(&mut self, path: &str)

pub fn add_translation(&mut self, lang: LanguageIdentifier, ftl: String)

pub fn spawn<F>(&self, target: F) where
F: 'static + Send + FnOnce(&mut ContextProxy),

pub fn process_data_updates(&mut self)

For each binding or data observer, check if its data has changed, and if so, rerun its builder/body.

pub fn process_style_updates(&mut self)

pub fn process_visual_updates(&mut self)

Massages the style system until everything is coherent

pub fn draw(&mut self, canvas: &mut Canvas<OpenGl>)

pub fn dispatch_system_event(&mut self, event: WindowEvent)

This method is in charge of receiving raw WindowEvents and dispatching them to the appropriate points in the tree.