Trait vizia::prelude::StyleModifiers

pub trait StyleModifiers: Modifiable {

    fn id(self, id: impl Into<String>) -> Self { ... }
    fn class(self, name: &str) -> Self { ... }
    fn toggle_class(self, name: &str, applied: impl Res<bool>) -> Self { ... }
    fn checked<U>(self, state: impl Res<U>) -> Self
    where
        U: Into<bool>,
    { ... }
    fn disabled<U>(self, value: impl Res<U>) -> Self
    where
        U: Into<bool>,
    { ... }
    fn display<U>(self, value: impl Res<U>) -> Self
    where
        U: Into<Display>,
    { ... }
    fn visibility<U>(self, value: impl Res<U>) -> Self
    where
        U: Into<Visibility>,
    { ... }
    fn z_order<U>(self, value: impl Res<U>) -> Self
    where
        U: Into<i32>,
    { ... }
    fn overflow<U>(self, value: impl Res<U>) -> Self
    where
        U: Into<Overflow>,
    { ... }
    fn background_color<U>(self, value: impl Res<U>) -> Self
    where
        U: Into<Color>,
    { ... }
    fn background_image<U>(self, value: impl Res<U>) -> Self
    where
        U: Into<String>,
    { ... }
    fn image<U>(self, value: impl Res<U>) -> Self
    where
        U: ToString,
    { ... }
    fn border_width<U>(self, value: impl Res<U>) -> Self
    where
        U: Into<Units>,
    { ... }
    fn border_color<U>(self, value: impl Res<U>) -> Self
    where
        U: Into<Color>,
    { ... }
    fn border_radius_top_left<U>(self, value: impl Res<U>) -> Self
    where
        U: Into<Units>,
    { ... }
    fn border_radius_top_right<U>(self, value: impl Res<U>) -> Self
    where
        U: Into<Units>,
    { ... }
    fn border_radius_bottom_left<U>(self, value: impl Res<U>) -> Self
    where
        U: Into<Units>,
    { ... }
    fn border_radius_bottom_right<U>(self, value: impl Res<U>) -> Self
    where
        U: Into<Units>,
    { ... }
    fn border_radius<U>(self, value: impl Res<U>) -> Self
    where
        U: Into<Units>,
    { ... }
    fn border_shape_top_left<U>(self, value: impl Res<U>) -> Self
    where
        U: Into<BorderCornerShape>,
    { ... }
    fn border_shape_top_right<U>(self, value: impl Res<U>) -> Self
    where
        U: Into<BorderCornerShape>,
    { ... }
    fn border_shape_bottom_left<U>(self, value: impl Res<U>) -> Self
    where
        U: Into<BorderCornerShape>,
    { ... }
    fn border_shape_bottom_right<U>(self, value: impl Res<U>) -> Self
    where
        U: Into<BorderCornerShape>,
    { ... }
    fn border_corner_shape<U>(self, value: impl Res<U>) -> Self
    where
        U: Into<BorderCornerShape>,
    { ... }
    fn outline_width<U>(self, value: impl Res<U>) -> Self
    where
        U: Into<Units>,
    { ... }
    fn outline_color<U>(self, value: impl Res<U>) -> Self
    where
        U: Into<Color>,
    { ... }
    fn outline_offset<U>(self, value: impl Res<U>) -> Self
    where
        U: Into<Units>,
    { ... }
    fn cursor<U>(self, value: impl Res<U>) -> Self
    where
        U: Into<CursorIcon>,
    { ... }
    fn rotate<U>(self, value: impl Res<U>) -> Self
    where
        U: Into<f32>,
    { ... }
    fn translate<U>(self, value: impl Res<U>) -> Self
    where
        U: Into<(f32, f32)>,
    { ... }
    fn scale<U>(self, value: impl Res<U>) -> Self
    where
        U: Into<(f32, f32)>,
    { ... }

}

Modifiers for changing the style properties of a view.

Provided Methods

fn id(self, id: impl Into<String>) -> Self

Sets the ID name of the view.

The ID name can be references by a CSS selector.

Example

Element::new(cx).id("foo");

css

#foo {
    background-color: red;
}

fn class(self, name: &str) -> Self

Adds a class name to the view.

fn toggle_class(self, name: &str, applied: impl Res<bool>) -> Self

Sets whether a view should have the given class name.

fn checked<U>(self, state: impl Res<U>) -> Selfwhere
U: Into<bool>,

Sets the state of the view to checked.

fn disabled<U>(self, value: impl Res<U>) -> Selfwhere
U: Into<bool>,

Sets the view to be disabled.

This property is inherited by the descendants of the view.

fn display<U>(self, value: impl Res<U>) -> Selfwhere
U: Into<Display>,

Sets whether the view should be positioned and rendered.

A display value of Display::None causes the view to be ignored by both layout and rendering.

fn visibility<U>(self, value: impl Res<U>) -> Selfwhere
U: Into<Visibility>,

Sets whether the view should be rendered.

The layout system will still compute the size and position of an invisible view.

fn z_order<U>(self, value: impl Res<U>) -> Selfwhere
U: Into<i32>,

Sets the z-order index of the view.

Views with a higher z-order will be rendered on top of those with a lower z-order. Views with the same z-order are rendered in tree order.

fn overflow<U>(self, value: impl Res<U>) -> Selfwhere
U: Into<Overflow>,

Sets the overflow behavior of the view.

The overflow behavior determines whether child views can render outside the bounds of their parent.

fn background_color<U>(self, value: impl Res<U>) -> Selfwhere
U: Into<Color>,

Sets the background color of the view.

fn background_image<U>(self, value: impl Res<U>) -> Selfwhere
U: Into<String>,

Sets the background image of the view.

Background image will override any background gradient or color.

fn image<U>(self, value: impl Res<U>) -> Selfwhere
U: ToString,

fn border_width<U>(self, value: impl Res<U>) -> Selfwhere
U: Into<Units>,

Sets the border width of the view.

fn border_color<U>(self, value: impl Res<U>) -> Selfwhere
U: Into<Color>,

Sets the border color of the view.

fn border_radius_top_left<U>(self, value: impl Res<U>) -> Selfwhere
U: Into<Units>,

Sets the border radius for the top-left corner of the view.

fn border_radius_top_right<U>(self, value: impl Res<U>) -> Selfwhere
U: Into<Units>,

Sets the border radius for the top-right corner of the view.

fn border_radius_bottom_left<U>(self, value: impl Res<U>) -> Selfwhere
U: Into<Units>,

Sets the border radius for the bottom-left corner of the view.

fn border_radius_bottom_right<U>(self, value: impl Res<U>) -> Selfwhere
U: Into<Units>,

Sets the border radius for the bottom-right corner of the view.

fn border_radius<U>(self, value: impl Res<U>) -> Selfwhere
U: Into<Units>,

Sets the border radius for all four corners of the view.

fn border_shape_top_left<U>(self, value: impl Res<U>) -> Selfwhere
U: Into<BorderCornerShape>,

Sets the border corner shape for the top-left corner of the view.

fn border_shape_top_right<U>(self, value: impl Res<U>) -> Selfwhere
U: Into<BorderCornerShape>,

Sets the border corner shape for the top-right corner of the view.

fn border_shape_bottom_left<U>(self, value: impl Res<U>) -> Selfwhere
U: Into<BorderCornerShape>,

Sets the border corner shape for the bottom-left corner of the view.

fn border_shape_bottom_right<U>(self, value: impl Res<U>) -> Selfwhere
U: Into<BorderCornerShape>,

Sets the border corner shape for the bottom-right corner of the view.

fn border_corner_shape<U>(self, value: impl Res<U>) -> Selfwhere
U: Into<BorderCornerShape>,

Sets the border corner shape for all four corners of the view.

fn outline_width<U>(self, value: impl Res<U>) -> Selfwhere
U: Into<Units>,

Sets the outline width of the view.

fn outline_color<U>(self, value: impl Res<U>) -> Selfwhere
U: Into<Color>,

Sets the outline color of the view.

fn outline_offset<U>(self, value: impl Res<U>) -> Selfwhere
U: Into<Units>,

Sets the outline offset of the view.

fn cursor<U>(self, value: impl Res<U>) -> Selfwhere
U: Into<CursorIcon>,

Sets the mouse cursor used when the view is hovered.

fn rotate<U>(self, value: impl Res<U>) -> Selfwhere
U: Into<f32>,

Sets the angle of rotation for the view.

Rotation applies to the rendered view and does not affect layout.

fn translate<U>(self, value: impl Res<U>) -> Selfwhere
U: Into<(f32, f32)>,

Sets the translation offset of the view.

Translation applies to the rendered view and does not affect layout.

fn scale<U>(self, value: impl Res<U>) -> Selfwhere
U: Into<(f32, f32)>,

Sets the scale of the view.

Scale applies to the rendered view and does not affect layout.