Trait vizia::prelude::View

pub trait View: 'static {
    fn build<F>(self, cx: &mut Context, content: F) -> Handle<'_, Self>
    where
        F: FnOnce(&mut Context),
    { ... }
    fn element(&self) -> Option<&'static str> { ... }
    fn event(&mut self, cx: &mut EventContext<'_>, event: &mut Event) { ... }
    fn draw(&self, cx: &mut DrawContext<'_>, canvas: &mut Canvas<OpenGl>) { ... }
}

A view is any object which can be displayed on the screen.

This trait is part of the prelude.

Provided Methods

fn build<F>(self, cx: &mut Context, content: F) -> Handle<'_, Self> where
F: FnOnce(&mut Context),

fn element(&self) -> Option<&'static str>

The name of the view. This is used in css: to style every single one of a given view, you specify the element name.

fn event(&mut self, cx: &mut EventContext<'_>, event: &mut Event)

fn draw(&self, cx: &mut DrawContext<'_>, canvas: &mut Canvas<OpenGl>)