vizia::view
Struct Handle

pub struct Handle<'a, V> { /* private fields */ }

A handle to a view which has been built into the tree.

Implementations

impl<V> Handle<'_, V>

pub fn entity(&self) -> Entity

Returns the [Entity] id of the view.

pub fn context(&mut self) -> &mut Context

Returns a mutable reference to the context.

pub fn parent(&self) -> Entity

Returns the entity id of the parent view.

pub fn lock_focus_to_within(self) -> Handle<'_, V>

Stop the user from tabbing out of a subtree, which is useful for modal dialogs.

pub fn modify<F>(self, f: F) -> Handle<'_, V>
where F: FnOnce(&mut V), V: 'static,

Mody the internal data of the view.

pub fn on_build<F>(self, callback: F) -> Handle<'_, V>
where F: Fn(&mut EventContext<'_>),

Callback which is run when the view is built/rebuilt.

pub fn bind<R, T, F>(self, res: R, closure: F) -> Handle<'_, V>
where R: Res<T>, F: 'static + Fn(Handle<'_, V>, R),

Creates a binding to the given lens and provides a closure which can be used to mutate the view through a handle.

pub fn needs_relayout(&mut self)

Marks the view as needing a relayout.

pub fn needs_restyle(&mut self)

Marks the view as needing a restyle.

pub fn needs_redraw(&mut self)

Marks the view as needing a redraw.

pub fn bounds(&self) -> BoundingBox

Returns the bounding box of the view.

pub fn scale_factor(&self) -> f32

Returns the scale factor of the device.

impl Handle<'_, Avatar>

pub fn variant<U>(self, variant: impl Res<U>) -> Handle<'_, Avatar>
where U: Into<AvatarVariant>,

Selects the geometric variant of the Avatar. Accepts a value of, or lens to, an AvatarVariant.

Avatar::new(cx, |cx|{
    Svg::new(cx, ICON_USER);
})
.variant(AvatarVariant::Rounded);

pub fn badge<F>(self, content: F) -> Handle<'_, Avatar>
where F: FnOnce(&mut Context) -> Handle<'_, Badge>,

Adds a badge to the Avatar.

Avatar::new(cx, |cx|{
    Svg::new(cx, ICON_USER);
})
.badge(|cx| Badge::empty(cx).class("error"));

impl Handle<'_, Badge>

pub fn placement<U>(self, placement: impl Res<U>) -> Handle<'_, Badge>
where U: Into<BadgePlacement>,

Sets the placement of a badge relative to its parent. Accepts a value of, or lens to, a BadgePlacement.

impl Handle<'_, ButtonGroup>

pub fn vertical(self, is_vertical: impl Res<bool>) -> Handle<'_, ButtonGroup>

Sets whether the button group is in vertical orientation.

impl Handle<'_, Checkbox>

pub fn on_toggle<F>(self, callback: F) -> Handle<'_, Checkbox>
where F: 'static + Fn(&mut EventContext<'_>),

Set the callback triggered when the checkbox is pressed.

Examples

Checkbox::new(cx, AppData::value)
    .on_toggle(|cx| cx.emit(AppEvent::ToggleValue));

impl Handle<'_, Chip>

pub fn on_close( self, callback: impl Fn(&mut EventContext<'_>) + Send + Sync + 'static, ) -> Handle<'_, Chip>

Set the callback triggered when the close button of the chip is pressed. The chip close button is not displayed by default. Setting this callback causes the close button to be displayed.

pub fn variant<U>(self, variant: impl Res<U>) -> Handle<'_, Chip>
where U: Into<ChipVariant>,

Selects the style variant to be used by the chip. Accepts a value of, or lens to, a ChipVariant.

Example

Chip::new(cx, "Chip")
    .variant(ChipVariant::Filled);

impl<L1, L2, T> Handle<'_, ComboBox<L1, L2, T>>
where L1: Lens<Target = Vec<T>>, T: 'static + Data + ToString, L2: Lens<Target = usize>,

pub fn on_select<F>(self, callback: F) -> Handle<'_, ComboBox<L1, L2, T>>
where F: 'static + Fn(&mut EventContext<'_>, usize),

Set the callback triggered when an item is selected from the ComboBox dropdown list.

impl Handle<'_, Datepicker>

pub fn on_select<F>(self, callback: F) -> Handle<'_, Datepicker>
where F: 'static + Fn(&mut EventContext<'_>, NaiveDate),

Set the callback triggered when a date is selected from the Datepicker view.

impl Handle<'_, Divider>

pub fn orientation( self, orientation: impl Res<Orientation>, ) -> Handle<'_, Divider>

Set the orientation of the divider. Accepts a value or a lens to an Orientation.

impl Handle<'_, Dropdown>

pub fn placement(self, placement: impl Res<Placement>) -> Handle<'_, Dropdown>

Sets the position where the tooltip should appear relative to its parent element. Defaults to Placement::Bottom.

pub fn show_arrow(self, show_arrow: bool) -> Handle<'_, Dropdown>

Sets whether the popup should include an arrow. Defaults to true.

pub fn arrow_size(self, size: impl Into<Length>) -> Handle<'_, Dropdown>

Sets the size of the popup arrow, or gap if the arrow is hidden.

pub fn should_reposition(self, flag: bool) -> Handle<'_, Dropdown>

Set to whether the popup should reposition to always be visible.

impl<L> Handle<'_, Knob<L>>
where L: Lens<Target = f32>,

pub fn on_change<F>(self, callback: F) -> Handle<'_, Knob<L>>
where F: 'static + Fn(&mut EventContext<'_>, f32),

Sets the callback triggered when the knob value is changed.

impl Handle<'_, ArcTrack>

pub fn value<L>(self, lens: L) -> Handle<'_, ArcTrack>
where L: Lens<Target = f32>,

impl Handle<'_, TickKnob>

pub fn value<L>(self, lens: L) -> Handle<'_, TickKnob>
where L: Lens<Target = f32>,

impl Handle<'_, Label>

pub fn describing( self, entity_identifier: impl Into<String>, ) -> Handle<'_, Label>

Which form element does this label describe.

Examples

Checkbox::new(cx, AppData::value).on_toggle(|cx| cx.emit(AppEvent::ToggleValue)).id("checkbox_identifier");
Label::new(cx, "hello").describing("checkbox_identifier");

impl Handle<'_, List>

pub fn selected<S>(self, selected: S) -> Handle<'_, List>
where S: Lens, <S as Lens>::Target: Deref<Target = [usize]> + Data,

Sets the selected items of the list. Takes a lens to a list of indices.

pub fn on_select<F>(self, callback: F) -> Handle<'_, List>
where F: 'static + Fn(&mut EventContext<'_>, usize),

Sets the callback triggered when a ListItem is selected.

pub fn selectable<U>(self, selectable: impl Res<U>) -> Handle<'_, List>
where U: Into<Selectable>,

Set the selectable state of the List.

pub fn selection_follows_focus<U>(self, flag: impl Res<U>) -> Handle<'_, List>
where U: Into<bool>,

Sets whether the selection should follow the focus.

pub fn horizontal<U>(self, flag: impl Res<U>) -> Handle<'_, List>
where U: Into<bool>,

Sets the orientation of the list.

impl Handle<'_, PickList>

pub fn placeholder<P>(self, placeholder: impl Res<P>) -> Handle<'_, PickList>
where P: ToStringLocalized,

Sets the placeholder text that appears when the textbox has no value.

pub fn on_select<F>(self, callback: F) -> Handle<'_, PickList>
where F: 'static + Fn(&mut EventContext<'_>, usize),

Sets the callback triggered when an option is selected.

impl Handle<'_, Popup>

pub fn placement(self, placement: impl Res<Placement>) -> Handle<'_, Popup>

Sets the position where the popup should appear relative to its parent element. Defaults to Placement::Bottom.

pub fn show_arrow(self, show_arrow: impl Res<bool>) -> Handle<'_, Popup>

Sets whether the popup should include an arrow. Defaults to true.

pub fn arrow_size<U>(self, size: impl Res<U>) -> Handle<'_, Popup>
where U: Into<Length>,

Sets the size of the popup arrow, or gap if the arrow is hidden.

pub fn should_reposition( self, should_reposition: impl Res<bool>, ) -> Handle<'_, Popup>

Set to whether the popup should reposition to always be visible.

pub fn on_blur<F>(self, f: F) -> Handle<'_, Popup>
where F: 'static + Fn(&mut EventContext<'_>),

Registers a callback for when the user clicks off of the popup, usually with the intent of closing it.

impl Handle<'_, RadioButton>

pub fn on_select<F>(self, callback: F) -> Handle<'_, RadioButton>
where F: 'static + Fn(&mut EventContext<'_>),

Set the callback triggered when the radio button is selected.

Examples

RadioButton::new(cx, AppData::value)
    .on_select(|cx| cx.emit(AppEvent::ToggleValue));

impl Handle<'_, Rating>

pub fn on_change<F>(self, callback: F) -> Handle<'_, Rating>
where F: 'static + Fn(&mut EventContext<'_>, u32),

Set the callback which is triggered when the rating changes.

impl<L1> Handle<'_, Scrollbar<L1>>
where L1: 'static + Lens<Target = f32>,

pub fn scroll_to_cursor( self, scroll_to_cursor: impl Res<bool>, ) -> Handle<'_, Scrollbar<L1>>

Sets whether the scrollbar should move to the cursor when pressed.

impl Handle<'_, ScrollView>

pub fn on_scroll( self, callback: impl Fn(&mut EventContext<'_>, f32, f32) + Send + Sync + 'static, ) -> Handle<'_, ScrollView>

Sets a callback which will be called when a scrollview is scrolled, either with the mouse wheel, touchpad, or using the scroll bars.

pub fn scroll_to_cursor(self, scroll_to_cursor: bool) -> Handle<'_, ScrollView>

Sets whether the scrollbar should move to the cursor when pressed.

pub fn scroll_x(self, scrollx: impl Res<f32>) -> Handle<'_, ScrollView>

Set the horizontal scroll position of the ScrollView. Accepts a value or lens to an ‘f32’ between 0 and 1.

pub fn scroll_y(self, scrollx: impl Res<f32>) -> Handle<'_, ScrollView>

Set the vertical scroll position of the ScrollView. Accepts a value or lens to an ‘f32’ between 0 and 1.

pub fn show_horizontal_scrollbar( self, flag: impl Res<bool>, ) -> Handle<'_, ScrollView>

Sets whether the horizontal scrollbar should be visible.

pub fn show_vertical_scrollbar( self, flag: impl Res<bool>, ) -> Handle<'_, ScrollView>

Sets whether the vertical scrollbar should be visible.

impl<L> Handle<'_, Slider<L>>
where L: Lens,

pub fn on_change<F>(self, callback: F) -> Handle<'_, Slider<L>>
where F: 'static + Fn(&mut EventContext<'_>, f32),

Sets the callback triggered when the slider value is changed.

Takes a closure which triggers when the slider value is changed, either by pressing the track or dragging the thumb along the track.

Example

Slider::new(cx, AppData::value)
    .on_change(|cx, value| {
        debug!("Slider on_changing: {}", value);
    });

pub fn range(self, range: Range<f32>) -> Handle<'_, Slider<L>>

Sets the range of the slider.

If the bound data is outside of the range then the slider will clip to min/max of the range.

Example

Slider::new(cx, AppData::value)
    .range(-20.0..50.0)
    .on_changing(|cx, value| {
        debug!("Slider on_changing: {}", value);
    });

pub fn step(self, step: f32) -> Handle<'_, Slider<L>>

Set the step value for the slider.

pub fn keyboard_fraction(self, keyboard_fraction: f32) -> Handle<'_, Slider<L>>

Sets the fraction of a slider that a press of an arrow key will change.

Example

Slider::new(cx, AppData::value)
    .keyboard_fraction(0.05)
    .on_changing(|cx, value| {
        debug!("Slider on_changing: {}", value);
    });

impl Handle<'_, Spinbox>

pub fn on_increment<F>(self, callback: F) -> Handle<'_, Spinbox>
where F: 'static + Fn(&mut EventContext<'_>) + Send + Sync,

Sets the callback which is triggered when the Spinbox value is incremented.

pub fn on_decrement<F>(self, callback: F) -> Handle<'_, Spinbox>
where F: 'static + Fn(&mut EventContext<'_>) + Send + Sync,

Sets the callback which is triggered when the Spinbox value is decremented.

pub fn orientation( self, orientation: impl Res<Orientation>, ) -> Handle<'_, Spinbox>

Sets the orientation of the Spinbox.

pub fn icons(self, icons: impl Res<SpinboxIcons>) -> Handle<'_, Spinbox>

Set the icons which should be used for the increment and decrement buttons of the Spinbox

impl Handle<'_, Switch>

pub fn on_toggle<F>(self, callback: F) -> Handle<'_, Switch>
where F: 'static + Fn(&mut EventContext<'_>),

Set the callback triggered when the Switch is pressed.

Examples

Switch::new(cx, AppData::value)
    .on_toggle(|cx| cx.emit(AppEvent::ToggleValue));

impl Handle<'_, TabView>

pub fn vertical(self) -> Handle<'_, TabView>

pub fn on_select( self, callback: impl Fn(&mut EventContext<'_>, usize) + 'static, ) -> Handle<'_, TabView>

pub fn with_selected<U>(self, selected: impl Res<U>) -> Handle<'_, TabView>
where U: Into<usize>,

impl<L> Handle<'_, Textbox<L>>
where L: Lens,

pub fn on_edit<F>(self, callback: F) -> Handle<'_, Textbox<L>>
where F: 'static + Fn(&mut EventContext<'_>, String) + Send + Sync,

Sets the callback triggered when a textbox is edited, i.e. text is inserted/deleted.

Callback provides the current text of the textbox.

pub fn on_submit<F>(self, callback: F) -> Handle<'_, Textbox<L>>
where F: 'static + Fn(&mut EventContext<'_>, <L as Lens>::Target, bool) + Send + Sync,

Sets the callback triggered when a textbox is submitted, i.e. when the enter key is pressed with a single-line textbox or the textbox loses focus.

Callback provides the text of the textbox and a flag to indicate if the submit was due to a key press or a loss of focus.

pub fn on_blur<F>(self, callback: F) -> Handle<'_, Textbox<L>>
where F: 'static + Fn(&mut EventContext<'_>) + Send + Sync,

Sets the callback triggered when a textbox is blurred, i.e. the mouse is pressed outside of the textbox.

pub fn on_cancel<F>(self, callback: F) -> Handle<'_, Textbox<L>>
where F: 'static + Fn(&mut EventContext<'_>) + Send + Sync,

Sets the callback triggered when a textbox edit is cancelled, i.e. the escape key is pressed while editing.

pub fn validate<F>(self, is_valid: F) -> Handle<'_, Textbox<L>>
where F: 'static + Fn(&<L as Lens>::Target) -> bool + Send + Sync,

Sets a validation closure which is called when the textbox is edited and sets the validity attribute to the output of the closure.

If a textbox is modified with the validate modifier then the on_submit will not be called if the text is invalid.

pub fn placeholder<P>(self, text: impl Res<P>) -> Handle<'_, Textbox<L>>
where P: ToStringLocalized,

Sets the placeholder text that appears when the textbox has no value.

impl Handle<'_, ToggleButton>

pub fn on_toggle( self, callback: impl Fn(&mut EventContext<'_>) + 'static, ) -> Handle<'_, ToggleButton>

Sets the callback triggered when the ToggleButton is toggled.

impl Handle<'_, Tooltip>

pub fn placement<U>(self, placement: impl Res<U>) -> Handle<'_, Tooltip>
where U: Into<Placement>,

Sets the position where the tooltip should appear relative to its parent element. Defaults to Placement::Bottom.

pub fn arrow<U>(self, show_arrow: impl Res<U>) -> Handle<'_, Tooltip>
where U: Into<bool>,

Sets whether the tooltip should include an arrow. Defaults to true.

pub fn arrow_size<U>(self, size: impl Res<U>) -> Handle<'_, Tooltip>
where U: Into<Length>,

Sets the size of the tooltip arrow if enabled.

impl Handle<'_, VirtualList>

pub fn scroll_to_cursor(self, flag: bool) -> Handle<'_, VirtualList>

Sets whether the scrollbar should move to the cursor when pressed.

impl Handle<'_, XYPad>

pub fn on_change<F>(self, callback: F) -> Handle<'_, XYPad>
where F: Fn(&mut EventContext<'_>, f32, f32) + 'static,

Set the callback which will be triggered when the XYPad is manipulated.