Struct vizia::view::Handle

pub struct Handle<'a, V> { /* private fields */ }

A handle to a view which has been built into the tree.

Implementations

impl<'a, V> Handle<'a, V>

pub fn entity(&self) -> Entity

Returns the [Entity] id of the view.

pub fn context(&mut self) -> &mut Context

Returns a mutable reference to the context.

pub fn parent(&self) -> Entity

pub fn lock_focus_to_within(self) -> Handle<'a, V>

Stop the user from tabbing out of a subtree, which is useful for modal dialogs.

pub fn modify<F>(self, f: F) -> Handle<'a, V>
where F: FnOnce(&mut V), V: 'static,

Mody the internal data of the view.

pub fn on_build<F>(self, callback: F) -> Handle<'a, V>
where F: Fn(&mut EventContext<'_>),

Callback which is run when the view is built/rebuilt.

pub fn bind<R, T, F>(self, res: R, closure: F) -> Handle<'a, V>
where R: Res<T>, F: 'static + Clone + Fn(Handle<'_, V>, R),

pub fn needs_relayout(&mut self)

Marks the view as needing a relayout.

pub fn needs_restyle(&mut self)

Marks the view as needing a restyle.

pub fn needs_redraw(&mut self)

Marks the view as needing a redraw.

pub fn bounds(&self) -> BoundingBox

Returns the bounding box of the view.

pub fn scale_factor(&self) -> f32

Returns the scale factor of the device.

impl<'a> Handle<'a, Avatar>

pub fn variant<U>(self, variant: impl Res<U>) -> Handle<'a, Avatar>
where U: Into<AvatarVariant>,

Selects the style variant for the Avatar.

pub fn badge<F>(self, content: F) -> Handle<'a, Avatar>
where F: FnOnce(&mut Context) -> Handle<'_, Badge>,

Adds a badge to the Avatar.

impl<'a> Handle<'a, Badge>

pub fn placement(self, placement: BadgePlacement) -> Handle<'a, Badge>

Sets the placement of a badge relative to its parent when used with the badge modifier.

impl<'a> Handle<'a, ButtonGroup>

pub fn vertical(self, is_vertical: impl Res<bool>) -> Handle<'a, ButtonGroup>

impl Handle<'_, Checkbox>

pub fn on_toggle<F>(self, callback: F) -> Handle<'_, Checkbox>
where F: 'static + Fn(&mut EventContext<'_>),

Set the callback triggered when the checkbox is pressed.

Examples

Checkbox::new(cx, AppData::value).on_toggle(|cx| cx.emit(AppEvent::ToggleValue));

impl<'a> Handle<'a, Chip>

pub fn on_close( self, callback: impl Fn(&mut EventContext<'_>) + Send + Sync + 'static, ) -> Handle<'a, Chip>

pub fn variant<U>(self, variant: impl Res<U>) -> Handle<'a, Chip>
where U: Into<ChipVariant>,

Selects the style variant to be used by the chip.

Example

Chip::new(cx, "Chip")
    .variant(ChipVariant::Filled);

impl<'v, L1, L2, T> Handle<'v, ComboBox<L1, L2, T>>
where L1: Lens<Target = Vec<T>>, T: 'static + Data + ToString, L2: Lens<Target = usize>,

pub fn on_select<F>(self, callback: F) -> Handle<'v, ComboBox<L1, L2, T>>
where F: 'static + Fn(&mut EventContext<'_>, usize),

impl<'a> Handle<'a, Datepicker>

pub fn on_select<F>(self, callback: F) -> Handle<'a, Datepicker>
where F: 'static + Fn(&mut EventContext<'_>, NaiveDate),

impl<'a> Handle<'a, Divider>

pub fn orientation( self, orientation: impl Res<Orientation>, ) -> Handle<'a, Divider>

impl<'a, L> Handle<'a, Knob<L>>
where L: Lens<Target = f32>,

pub fn on_changing<F>(self, callback: F) -> Handle<'a, Knob<L>>
where F: 'static + Fn(&mut EventContext<'_>, f32),

impl Handle<'_, ArcTrack>

pub fn value<L>(self, lens: L) -> Handle<'_, ArcTrack>
where L: Lens<Target = f32>,

impl Handle<'_, TickKnob>

pub fn value<L>(self, lens: L) -> Handle<'_, TickKnob>
where L: Lens<Target = f32>,

impl Handle<'_, Label>

pub fn describing( self, entity_identifier: impl Into<String>, ) -> Handle<'_, Label>

Which form element does this label describe.

Examples

Checkbox::new(cx, AppData::value).on_toggle(|cx| cx.emit(AppEvent::ToggleValue)).id("checkbox_identifier");
Label::new(cx, "hello").describing("checkbox_identifier");

impl<'v> Handle<'v, PickList>

pub fn on_select<F>(self, callback: F) -> Handle<'v, PickList>
where F: 'static + Fn(&mut EventContext<'_>, usize),

impl<'v> Handle<'v, ScrollList>

pub fn on_select<F>(self, callback: F) -> Handle<'v, ScrollList>
where F: 'static + Fn(&mut EventContext<'_>, usize),

impl<'a> Handle<'a, Popup>

pub fn placement(self, placement: impl Res<Placement>) -> Handle<'a, Popup>

Sets the position where the tooltip should appear relative to its parent element. Defaults to Placement::Bottom.

pub fn arrow(self, show_arrow: bool) -> Handle<'a, Popup>

Sets whether the popup should include an arrow. Defaults to true.

pub fn arrow_size(self, size: impl Into<Length>) -> Handle<'a, Popup>

Sets the size of the popup arrow, or gap if the arrow is hidden.

pub fn should_reposition(self, flag: bool) -> Handle<'a, Popup>

Set to whether the popup should reposition to always be visible.

pub fn on_blur<F>(self, f: F) -> Handle<'a, Popup>
where F: 'static + Fn(&mut EventContext<'_>),

Registers a callback for when the user clicks off of the popup, usually with the intent of closing it.

impl<'a> Handle<'a, ProgressBar>

pub fn bar_color(self, color: impl Res<Color>) -> Handle<'a, ProgressBar>

Set the color of the bar inside the ProgressBar.

you also pass a lens to this method if you want to be able to change the color dynamically.

impl Handle<'_, RadioButton>

pub fn on_select<F>(self, callback: F) -> Handle<'_, RadioButton>
where F: 'static + Fn(&mut EventContext<'_>),

Set the callback triggered when the radio button is selected.

Examples

RadioButton::new(cx, AppData::value).on_select(|cx| cx.emit(AppEvent::ToggleValue));

impl<'a> Handle<'a, Rating>

pub fn on_change<F>(self, callback: F) -> Handle<'a, Rating>
where F: 'static + Fn(&mut EventContext<'_>, u32),

impl<'a, L1> Handle<'a, Scrollbar<L1>>
where L1: 'static + Lens<Target = f32>,

pub fn scroll_to_cursor( self, scroll_to_cursor: impl Res<bool>, ) -> Handle<'a, Scrollbar<L1>>

impl<'a> Handle<'a, ScrollView>

pub fn on_scroll( self, callback: impl Fn(&mut EventContext<'_>, f32, f32) + Send + Sync + 'static, ) -> Handle<'a, ScrollView>

Sets a callback which will be called when a scrollview is scrolled, either with the mouse wheel, touchpad, or using the scroll bars.

pub fn scroll_to_cursor(self, scroll_to_cursor: bool) -> Handle<'a, ScrollView>

pub fn scrollx(self, scrollx: impl Res<f32>) -> Handle<'a, ScrollView>

pub fn scrolly(self, scrollx: impl Res<f32>) -> Handle<'a, ScrollView>

impl<L> Handle<'_, Slider<L>>
where L: Lens,

pub fn on_changing<F>(self, callback: F) -> Handle<'_, Slider<L>>
where F: 'static + Fn(&mut EventContext<'_>, f32),

Sets the callback triggered when the slider value is changing (dragging).

Takes a closure which triggers when the slider value is changing, either by pressing the track or dragging the thumb along the track.

Example

Slider::new(cx, AppData::value)
    .on_changing(|cx, value| {
        debug!("Slider on_changing: {}", value);
    });

pub fn range(self, range: Range<f32>) -> Handle<'_, Slider<L>>

Sets the range of the slider.

If the bound data is outside of the range then the slider will clip to min/max of the range.

Example

Slider::new(cx, AppData::value)
    .range(-20.0..50.0)
    .on_changing(|cx, value| {
        debug!("Slider on_changing: {}", value);
    });

pub fn step(self, step: f32) -> Handle<'_, Slider<L>>

pub fn keyboard_fraction(self, keyboard_fraction: f32) -> Handle<'_, Slider<L>>

Sets the fraction of a slider that a press of an arrow key will change.

Example

Slider::new(cx, AppData::value)
    .keyboard_fraction(0.05)
    .on_changing(|cx, value| {
        debug!("Slider on_changing: {}", value);
    });

impl Handle<'_, NamedSlider>

pub fn on_changing<F>(self, callback: F) -> Handle<'_, NamedSlider>
where F: 'static + Fn(&mut EventContext<'_>, f32),

pub fn range(self, range: Range<f32>) -> Handle<'_, NamedSlider>

pub fn keyboard_fraction( self, keyboard_fraction: f32, ) -> Handle<'_, NamedSlider>

impl<'a> Handle<'a, Spinbox>

pub fn on_increment<F>(self, callback: F) -> Handle<'a, Spinbox>
where F: 'static + Fn(&mut EventContext<'_>) + Send + Sync,

pub fn on_decrement<F>(self, callback: F) -> Handle<'a, Spinbox>
where F: 'static + Fn(&mut EventContext<'_>) + Send + Sync,

pub fn orientation( self, orientation: impl Res<Orientation>, ) -> Handle<'a, Spinbox>

pub fn icons(self, icons: impl Res<SpinboxIcons>) -> Handle<'a, Spinbox>

impl Handle<'_, Switch>

pub fn on_toggle<F>(self, callback: F) -> Handle<'_, Switch>
where F: 'static + Fn(&mut EventContext<'_>),

Set the callback triggered when the Switch is pressed.

Examples

Switch::new(cx, AppData::value).on_toggle(|cx| cx.emit(AppEvent::ToggleValue));

impl<'a> Handle<'a, TabView>

pub fn vertical(self) -> Handle<'a, TabView>

pub fn on_select( self, callback: impl Fn(&mut EventContext<'_>, usize) + 'static, ) -> Handle<'a, TabView>

pub fn with_selected<U>(self, selected: impl Res<U>) -> Handle<'a, TabView>
where U: Into<usize>,

impl<'a, L> Handle<'a, Textbox<L>>
where L: Lens,

pub fn on_edit<F>(self, callback: F) -> Handle<'a, Textbox<L>>
where F: 'static + Fn(&mut EventContext<'_>, String) + Send + Sync,

Sets the callback triggered when a textbox is edited, i.e. text is inserted/deleted.

Callback provides the current text of the textbox.

pub fn on_submit<F>(self, callback: F) -> Handle<'a, Textbox<L>>
where F: 'static + Fn(&mut EventContext<'_>, <L as Lens>::Target, bool) + Send + Sync,

Sets the callback triggered when a textbox is submitted, i.e. when the enter key is pressed with a single-line textbox or the textbox loses focus.

Callback provides the text of the textbox and a flag to indicate if the submit was due to a key press or a loss of focus.

pub fn on_blur<F>(self, callback: F) -> Handle<'a, Textbox<L>>
where F: 'static + Fn(&mut EventContext<'_>) + Send + Sync,

Sets the callback triggered when a textbox is blurred, i.e. the mouse is pressed outside of the textbox.

pub fn on_cancel<F>(self, callback: F) -> Handle<'a, Textbox<L>>
where F: 'static + Fn(&mut EventContext<'_>) + Send + Sync,

Sets the callback triggered when a textbox edit is cancelled, i.e. the escape key is pressed while editing.

pub fn validate<F>(self, is_valid: F) -> Handle<'a, Textbox<L>>
where F: 'static + Fn(&<L as Lens>::Target) -> bool + Send + Sync,

Sets a validation closure which is called when the textbox is edited and sets the validity attribute to the output of the closure.

If a textbox is modified with the validate modifier then the on_submit will not be called if the text is invalid.

pub fn placeholder<P>(self, text: impl Res<P>) -> Handle<'a, Textbox<L>>
where P: ToStringLocalized,

Sets the placeholder text that appears when the textbox has no value.

impl<'a> Handle<'a, Tooltip>

pub fn placement(self, placement: Placement) -> Handle<'a, Tooltip>

Sets the position where the tooltip should appear relative to its parent element. Defaults to Placement::Bottom.

pub fn arrow(self, show_arrow: bool) -> Handle<'a, Tooltip>

Sets whether the tooltip should include an arrow. Defaults to true.

pub fn arrow_size(self, size: impl Into<Length>) -> Handle<'a, Tooltip>

impl<'a> Handle<'a, VirtualList>

pub fn scroll_to_cursor(self, flag: bool) -> Handle<'a, VirtualList>

impl Handle<'_, XYPad>

pub fn on_change<F>(self, callback: F) -> Handle<'_, XYPad>
where F: Fn(&mut EventContext<'_>, f32, f32) + 'static,