Struct vizia::view::Handle

pub struct Handle<'a, V> { /* private fields */ }

A handle to a view which has been already built into the tree.

Implementations

impl<'a, V> Handle<'a, V>

pub fn entity(&self) -> Entity

Returns the [Entity] id of the view.

pub fn parent(&self) -> Entity

pub fn lock_focus_to_within(self) -> Handle<'a, V>

Stop the user from tabbing out of a subtree, which is useful for modal dialogs.

pub fn modify<F>(self, f: F) -> Handle<'a, V>where F: FnOnce(&mut V), V: 'static,

Mody the internal data of the view.

pub fn on_build<F>(self, callback: F) -> Handle<'a, V>where F: Fn(&mut EventContext<'_>),

Callback which is run when the view is built/rebuilt.

pub fn bind<L, F>(self, lens: L, closure: F) -> Handle<'a, V>where L: Lens, <L as Lens>::Target: Data, F: 'static + Fn(Handle<'_, V>, L),

pub fn needs_relayout(&mut self)

Marks the view as needing a relayout.

pub fn needs_restyle(&mut self)

Marks the view as needing a restyle.

pub fn needs_redraw(&mut self)

Marks the view as needing a redraw.

pub fn bounds(&self) -> BoundingBox

Returns the bounding box of the view.

pub fn scale_factor(&self) -> f32

Returns the scale factor of the device.

impl Handle<'_, Checkbox>

pub fn on_toggle<F>(self, callback: F) -> Handle<'_, Checkbox>where F: 'static + Fn(&mut EventContext<'_>),

Set the callback triggered when the checkbox is pressed.

Examples

Checkbox::new(cx, AppData::value).on_toggle(|cx| cx.emit(AppEvent::ToggleValue));

impl<'a> Handle<'a, Chip>

pub fn on_close( self, callback: impl Fn(&mut EventContext<'_>) + Send + Sync + 'static ) -> Handle<'a, Chip>

impl<'v, L1, L2, T> Handle<'v, ComboBox<L1, L2, T>>where L1: Lens<Target = Vec<T, Global>>, T: 'static + Data + ToString, L2: Lens<Target = usize>,

pub fn on_select<F>(self, callback: F) -> Handle<'v, ComboBox<L1, L2, T>>where F: 'static + Fn(&mut EventContext<'_>, usize),

impl<'a> Handle<'a, Datepicker>

pub fn on_select<F>(self, callback: F) -> Handle<'a, Datepicker>where F: 'static + Fn(&mut EventContext<'_>, NaiveDate),

impl<'a, L> Handle<'a, Knob<L>>where L: Lens<Target = f32>,

pub fn on_changing<F>(self, callback: F) -> Handle<'a, Knob<L>>where F: 'static + Fn(&mut EventContext<'_>, f32),

impl Handle<'_, ArcTrack>

pub fn value<L>(self, lens: L) -> Handle<'_, ArcTrack>where L: Lens<Target = f32>,

impl Handle<'_, TickKnob>

pub fn value<L>(self, lens: L) -> Handle<'_, TickKnob>where L: Lens<Target = f32>,

impl Handle<'_, Label>

pub fn describing( self, entity_identifier: impl Into<String> ) -> Handle<'_, Label>

Which form element does this label describe.

Examples

Checkbox::new(cx, AppData::value).on_toggle(|cx| cx.emit(AppEvent::ToggleValue)).id("checkbox_identifier");
Label::new(cx, "hello").describing("checkbox_identifier");

impl<L, T> Handle<'_, List<L, T>>where L: Lens<Target = Vec<T, Global>>, T: Data,

pub fn on_increment<F>(self, callback: F) -> Handle<'_, List<L, T>>where F: 'static + Fn(&mut EventContext<'_>),

pub fn on_decrement<F>(self, callback: F) -> Handle<'_, List<L, T>>where F: 'static + Fn(&mut EventContext<'_>),

pub fn on_clear<F>(self, callback: F) -> Handle<'_, List<L, T>>where F: 'static + Fn(&mut EventContext<'_>),

impl<'a, L> Handle<'a, MenuPopup<L>>where L: Lens, <L as Lens>::Target: Clone + Into<bool>,

pub fn on_blur<F>(self, f: F) -> Handle<'a, MenuPopup<L>>where F: 'static + Fn(&mut EventContext<'_>),

Registers a callback for when the user clicks off of the popup, usually with the intent of closing it.

impl<'v> Handle<'v, PickList>

pub fn on_select<F>(self, callback: F) -> Handle<'v, PickList>where F: 'static + Fn(&mut EventContext<'_>, usize),

impl<'a, L> Handle<'a, Popup<L>>where L: Lens, <L as Lens>::Target: Clone + Into<bool>,

pub fn on_blur<F>(self, f: F) -> Handle<'a, Popup<L>>where F: 'static + Fn(&mut EventContext<'_>),

Registers a callback for when the user clicks off of the popup, usually with the intent of closing it.

impl Handle<'_, RadioButton>

pub fn on_select<F>(self, callback: F) -> Handle<'_, RadioButton>where F: 'static + Fn(&mut EventContext<'_>),

Set the callback triggered when the radio button is selected.

Examples

RadioButton::new(cx, AppData::value).on_select(|cx| cx.emit(AppEvent::ToggleValue));

impl<'a> Handle<'a, Rating>

pub fn on_change<F>(self, callback: F) -> Handle<'a, Rating>where F: 'static + Fn(&mut EventContext<'_>, u32),

impl<'a, L> Handle<'a, ScrollView<L>>where L: Lens,

pub fn on_scroll( self, callback: impl Fn(&mut EventContext<'_>, f32, f32) + Send + Sync + 'static ) -> Handle<'a, ScrollView<L>>

impl<L> Handle<'_, Slider<L>>where L: Lens,

pub fn on_changing<F>(self, callback: F) -> Handle<'_, Slider<L>>where F: 'static + Fn(&mut EventContext<'_>, f32),

Sets the callback triggered when the slider value is changing (dragging).

Takes a closure which triggers when the slider value is changing, either by pressing the track or dragging the thumb along the track.

Example

Slider::new(cx, AppData::value)
    .on_changing(|cx, value| {
        println!("Slider on_changing: {}", value);
    });

pub fn range(self, range: Range<f32>) -> Handle<'_, Slider<L>>

Sets the range of the slider.

If the bound data is outside of the range then the slider will clip to min/max of the range.

Example

Slider::new(cx, AppData::value)
    .range(-20.0..50.0)
    .on_changing(|cx, value| {
        println!("Slider on_changing: {}", value);
    });

pub fn step(self, step: f32) -> Handle<'_, Slider<L>>

pub fn keyboard_fraction(self, keyboard_fraction: f32) -> Handle<'_, Slider<L>>

Sets the fraction of a slider that a press of an arrow key will change.

Example

Slider::new(cx, AppData::value)
    .keyboard_fraction(0.05)
    .on_changing(|cx, value| {
        println!("Slider on_changing: {}", value);
    });

impl Handle<'_, NamedSlider>

pub fn on_changing<F>(self, callback: F) -> Handle<'_, NamedSlider>where F: 'static + Fn(&mut EventContext<'_>, f32),

pub fn range(self, range: Range<f32>) -> Handle<'_, NamedSlider>

pub fn keyboard_fraction( self, keyboard_fraction: f32 ) -> Handle<'_, NamedSlider>

impl<'a> Handle<'a, Spinbox>

pub fn on_increment<F>(self, callback: F) -> Handle<'a, Spinbox>where F: 'static + Fn(&mut EventContext<'_>) + Send + Sync,

pub fn on_decrement<F>(self, callback: F) -> Handle<'a, Spinbox>where F: 'static + Fn(&mut EventContext<'_>) + Send + Sync,

impl Handle<'_, Switch>

pub fn on_toggle<F>(self, callback: F) -> Handle<'_, Switch>where F: 'static + Fn(&mut EventContext<'_>),

Set the callback triggered when the Switch is pressed.

Examples

Switch::new(cx, AppData::value).on_toggle(|cx| cx.emit(AppEvent::ToggleValue));

impl<'a, L> Handle<'a, Textbox<L>>where L: Lens,

pub fn on_edit<F>(self, callback: F) -> Handle<'a, Textbox<L>>where F: 'static + Fn(&mut EventContext<'_>, String) + Send + Sync,

Sets the callback triggered when a textbox is edited, i.e. text is inserted/deleted.

Callback provides the current text of the textbox.

pub fn on_submit<F>(self, callback: F) -> Handle<'a, Textbox<L>>where F: 'static + Fn(&mut EventContext<'_>, String, bool) + Send + Sync,

Sets the callback triggered when a textbox is submitted, i.e. when the enter key is pressed with a single-line textbox or the textbox loses focus.

Callback provides the text of the textbox and a flag to indicate if the submit was due to a key press or a loss of focus.

pub fn on_blur<F>(self, callback: F) -> Handle<'a, Textbox<L>>where F: 'static + Fn(&mut EventContext<'_>) + Send + Sync,

pub fn validate<F>(self, is_valid: F) -> Handle<'a, Textbox<L>>where F: 'static + Fn(&String) -> bool + Send + Sync,

Sets a validation closure which is called when the textbox is edited and sets the validity attribute to the output of the closure.

If a textbox is modified with the validate modifier then the on_submit will not be called if the text is invalid.

pub fn placeholder<T>(self, text: impl Res<T>) -> Handle<'a, Textbox<L>>where T: ToString,

impl<'a> Handle<'a, Timepicker>

pub fn on_change<F>(self, callback: F) -> Handle<'a, Timepicker>where F: 'static + Fn(&mut EventContext<'_>, NaiveTime),

impl<'a, L, T> Handle<'a, DigitalTimepicker<L, T>>where L: Lens<Target = T>, T: Timelike + Data,

pub fn on_change<F>(self, callback: F) -> Handle<'a, DigitalTimepicker<L, T>>where F: 'static + Fn(&mut EventContext<'_>, NaiveTime),

impl<'v, L, T> Handle<'v, AnalogTimepicker<L, T>>where L: Lens<Target = T>, T: Timelike + Data + Copy,

pub fn on_change<F>(self, callback: F) -> Handle<'v, AnalogTimepicker<L, T>>where F: 'static + Fn(&mut EventContext<'_>, NaiveTime),

pub fn show_controls(self, value: bool) -> Handle<'v, AnalogTimepicker<L, T>>

pub fn change_page_on_select( self, value: bool ) -> Handle<'v, AnalogTimepicker<L, T>>