Module vizia::views

Built-in views provided by vizia.

Modules

combo_box_derived_lenses

Derived lenses for ComboBox.

menu_bar_derived_lenses

Derived lenses for MenuBar.

menu_button_derived_lenses

Derived lenses for MenuButton.

normalized_map

popup_data_derived_lenses

Derived lenses for PopupData.

submenu_derived_lenses

Derived lenses for Submenu.

virtual_list_derived_lenses

Derived lenses for VirtualList.

Structs

AnalogTimepicker

ArcTrack

Makes a knob that represents the current value with an arc

Binding

A view with a binding which rebuilds its contents when the observed data changes.

Button

A simple push button with an action and a contained view.

Checkbox

A checkbox used to display and toggle a boolean state.

Chip

A visual indicator such as a tag.

ComboBox

ComboPopup

Datepicker

A control used to select a date.

DayTime

DigitalTimepicker

Dropdown

A dropdown is used to display some state with the ability to open a popup with options to change that state.

Element

A basic element with no interactivity.

HStack

A view which arranges its children into a horizontal stack (row).

Icon

Image

Knob

Label

A label used to display text.

List

A view for creating a list of items from a binding to a Vec<T>

MenuBar

MenuButton

MenuDivider

MenuPopup

NamedSlider

Notification

PickList

Popup

PopupData

ProgressBar

A simple ProgressBar that can be used to show progress of something.

RadioButton

A radio button used to display and toggle boolean state.

Rating

ScrollData

ScrollView

Scrollbar

Slider

The slider control can be used to select from a continuous set of values.

Spinbox

Submenu

Switch

A Switch used to display and toggle a boolean state.

TabPair

TabView

Textbox

The Textbox view provides an input control for editing a value as a string.

TickKnob

Makes a round knob with a tick to show the current value

Ticks

Adds tickmarks to a knob to show the steps that a knob can be set to. When added to a knob, the knob should be made smaller (depending on span), so the knob doesn’t overlap with the tick marks

Timepicker

Tooltip

VStack

A view which arranges its children into a vertical stack (column).

VirtualList

ZStack

A view which overlays its children on top of each other.

Enums

AMOrPM

AnalogTimepickerEvent

AnalogTimepickerPage

ComboBoxEvent

DigitalTimepickerEvent

KnobMode

MenuEvent

Orientation

The orientation of a widget, such as a slider or scrollbar

PopupEvent

ScrollEvent

SpinboxEvent

SpinboxIcons

SpinboxKind

TextEvent

Events for modifying a textbox.

VirtualListEvent